Module forest_filecoin::blocks::election_proof

Structs

EXP_DENO_COEF 🔒

EXP_NUM_COEF 🔒

ElectionProof

Proofs generated by a miner which determines the reward they earn. This is generated from hashing a partial ticket and using the hash to generate a value.

Poiss 🔒

Poisson inverted CDF lambda is in Q.256 format

Constants

MAX_WIN_COUNT 🔒

PRECISION 🔒

Functions

expneg 🔒

expneg accepts x in Q.256 format and computes e^-x. It is most precise within [0, 1.725) range, where error is less than 3.4e-30. Over the [0, 5) range its error is less than 4.6e-15. Output is in Q.256 format.

lambda 🔒

computes lambda in Q.256

parse 🔒

poly_val 🔒

poly_val evaluates a polynomial given by coefficients p in Q.256 format at point x in Q.256 format. Output is in Q.256. Coefficients should be ordered from the highest order coefficient to the lowest.