Module forest_filecoin::blocks::tipset

Modules

lotus_json

Tipset isn’t just plain old data - it has an invariant (all BlockHeaders are valid) So there is custom de-serialization here

Structs

FullTipset

FullTipset is an expanded version of a tipset that contains all the blocks and messages

Tipset

An immutable set of blocks at the same height with the same parent set. Blocks in a tipset are canonically ordered by ticket size.

TipsetKeys

A set of CIDs forming a unique key for a Tipset. Equal keys will have equivalent iteration order, but note that the CIDs are not maintained in the same order as the canonical iteration order of blocks in a tipset (which is by ticket)

Functions

verify_blocks 🔒