Module forest_filecoin::chain::store::chain_store

Modules

headchange_json

Structs

ChainStore

Stores chain data such as heaviest tipset and cached tipset info at each epoch. This structure is thread-safe, and all caches are wrapped in a mutex to allow a consistent ChainStore to be shared across tasks.

Enums

HeadChange

Enum for pubsub channel that defines message type variant and data contained in message type.

Constants

SINK_CAP 🔒

Functions

block_messages

Returns a Tuple of BLS messages of type UnsignedMessage and SECP messages of type SignedMessage

block_messages_from_cids

Returns a tuple of UnsignedMessage and SignedMessages from their CID

get_chain_message

Attempts to de-serialize to unsigned message or signed message and then returns it as a ChainMessage.

get_parent_receipt

Returns parent message receipt given block_header and message index.

messages_for_tipset

Given a tipset this function will return all unique messages in that tipset.

messages_from_cids

Returns messages from key-value store based on a slice of Cids.

persist_objects

Persists slice of serializable objects to blockstore.

read_amt_cids 🔒

Returns a vector of CIDs from provided root CID

read_msg_cids

Returns a tuple of CIDs for both unsigned and signed messages

Type Aliases

ChainEpochDelta

Disambiguate the type to signify that we are expecting a delta and not an actual epoch/height while maintaining the same type.