Struct forest_filecoin::chain_sync::chain_muxer::ChainMuxer

pub struct ChainMuxer<DB, M> {
    state: ChainMuxerState,
    worker_state: Arc<RwLock<SyncState>>,
    state_manager: Arc<StateManager<DB>>,
    network: SyncNetworkContext<DB>,
    genesis: Arc<Tipset>,
    bad_blocks: Arc<BadBlockCache>,
    net_handler: Receiver<NetworkEvent>,
    mpool: Arc<MessagePool<M>>,
    tipset_sender: Sender<Arc<Tipset>>,
    tipset_receiver: Receiver<Arc<Tipset>>,
    sync_config: SyncConfig,
    stateless_mode: bool,
}

The ChainMuxer handles events from the P2P network and orchestrates the chain synchronization.

Fields
state: ChainMuxerState

State of the ChainSyncer Future implementation

worker_state: Arc<RwLock<SyncState>>

Syncing state of chain sync workers.

state_manager: Arc<StateManager<DB>>

manages retrieving and updates state objects

network: SyncNetworkContext<DB>

Context to be able to send requests to P2P network

genesis: Arc<Tipset>

Genesis tipset

bad_blocks: Arc<BadBlockCache>

Bad blocks cache, updates based on invalid state transitions. Will mark any invalid blocks and all children as bad in this bounded cache

net_handler: Receiver<NetworkEvent>

Incoming network events to be handled by synchronizer

mpool: Arc<MessagePool<M>>

Message pool

tipset_sender: Sender<Arc<Tipset>>

Tipset channel sender

tipset_receiver: Receiver<Arc<Tipset>>

Tipset channel receiver

sync_config: SyncConfig

Syncing configurations

stateless_mode: bool

When stateless_mode is true, forest connects to the P2P network but does not sync to HEAD.

Implementations

impl<DB, M> ChainMuxer<DB, M>where DB: Blockstore + Sync + Send + 'static, M: Provider + Sync + Send + 'static,

pub fn new( state_manager: Arc<StateManager<DB>>, peer_manager: Arc<PeerManager>, mpool: Arc<MessagePool<M>>, network_send: Sender<NetworkMessage>, network_rx: Receiver<NetworkEvent>, genesis: Arc<Tipset>, tipset_sender: Sender<Arc<Tipset>>, tipset_receiver: Receiver<Arc<Tipset>>, cfg: SyncConfig, stateless_mode: bool ) -> Result<Self, ChainMuxerError>

pub fn bad_blocks_cloned(&self) -> Arc<BadBlockCache>

Returns a clone of the bad blocks cache to be used outside of chain sync.

pub fn sync_state_cloned(&self) -> Arc<RwLock<SyncState>>

Returns a cloned Arc of the sync worker state.

async fn get_full_tipset( network: SyncNetworkContext<DB>, chain_store: Arc<ChainStore<DB>>, peer_id: PeerId, tipset_keys: TipsetKeys ) -> Result<FullTipset, ChainMuxerError>

fn load_full_tipset( chain_store: Arc<ChainStore<DB>>, tipset_keys: TipsetKeys ) -> Result<FullTipset, ChainMuxerError>

async fn handle_peer_connected_event( network: SyncNetworkContext<DB>, chain_store: Arc<ChainStore<DB>>, peer_id: PeerId, genesis_block_cid: Cid )

async fn handle_peer_disconnected_event( network: SyncNetworkContext<DB>, peer_id: PeerId )

async fn gossipsub_block_to_full_tipset( block: GossipBlock, source: PeerId, network: SyncNetworkContext<DB> ) -> Result<FullTipset, ChainMuxerError>

fn handle_pubsub_message(mem_pool: Arc<MessagePool<M>>, message: SignedMessage)

async fn process_gossipsub_event( event: NetworkEvent, network: SyncNetworkContext<DB>, chain_store: Arc<ChainStore<DB>>, bad_block_cache: Arc<BadBlockCache>, mem_pool: Arc<MessagePool<M>>, genesis: Arc<Tipset>, message_processing_strategy: PubsubMessageProcessingStrategy, block_delay: u64 ) -> Result<Option<(FullTipset, PeerId)>, ChainMuxerError>

fn stateless_node( &self ) -> Pin<Box<dyn Future<Output = Result<(), ChainMuxerError>> + Send>>

fn evaluate_network_head( &self ) -> Pin<Box<dyn Future<Output = Result<NetworkHeadEvaluation, ChainMuxerError>> + Send>>

fn bootstrap( &self, network_head: FullTipset, local_head: Arc<Tipset> ) -> Pin<Box<dyn Future<Output = Result<(), ChainMuxerError>> + Send>>

fn follow( &self, tipset_opt: Option<FullTipset> ) -> Pin<Box<dyn Future<Output = Result<(), ChainMuxerError>> + Send>>