Function forest_filecoin::chain_sync::tipset_syncer::sync_messages_check_state

async fn sync_messages_check_state<DB: Blockstore + Send + Sync + 'static>(
    tracker: Arc<RwLock<SyncState>>,
    state_manager: Arc<StateManager<DB>>,
    network: SyncNetworkContext<DB>,
    chainstore: Arc<ChainStore<DB>>,
    bad_block_cache: &BadBlockCache,
    tipsets: Vec<Arc<Tipset>>,
    genesis: &Tipset,
    invalid_block_strategy: InvalidBlockStrategy
) -> Result<(), TipsetRangeSyncerError>

Going forward along the tipsets, try to load the messages in them from the BlockStore, or download them from the network, then validate the full tipset on each epoch.