Struct forest_filecoin::cid_collections::hash_map::CidHashMap

pub struct CidHashMap<V> {
    compact: HashMap<CidV1DagCborBlake2b256, V>,
    uncompact: HashMap<Uncompactable, V>,
}

A space-optimised hash map of [Cid]s, matching the API for std::collections::HashMap.

We accept the implementation complexity of per-compaction-method HashMaps for the space savings, which are constant per-variant, rather than constant per-item.

This is dramatic for large maps! Using, e.g frozen_vec::SmallCid will cost 25% more per-CID in the median case (32 B vs 40 B)

Fields
compact: HashMap<CidV1DagCborBlake2b256, V>uncompact: HashMap<Uncompactable, V>
Implementations

impl<V> CidHashMap<V>

pub fn new() -> Self

Creates an empty HashMap.

See also HashMap::new.

pub fn len(&self) -> usize

Returns the number of elements in the map.

See also HashMap::len.

pub fn capacity_min(&self) -> usize

How many values this map is guaranteed to hold without reallocating.

pub fn total_capacity(&self) -> usize

Reflective of reserved capacity of this map.

pub fn contains_key(&self, key: &Cid) -> bool

Returns true if the map contains a value for the specified key.

See also HashMap::contains_key.

pub fn get(&self, key: &Cid) -> Option<&V>

Returns a reference to the value corresponding to the key.

See also HashMap::get.

pub fn insert(&mut self, key: Cid, value: V) -> Option<V>

Inserts a key-value pair into the map.

If the map did not have this key present, None is returned.

If the map did have this key present, the value is updated, and the old value is returned.

See also HashMap::insert.

pub fn remove(&mut self, key: &Cid) -> Option<V>

Removes a key from the map, returning the value at the key if the key was previously in the map.

See also HashMap::remove.

impl<V> CidHashMap<V>

pub fn entry(&mut self, key: Cid) -> Entry<'_, V>

Gets the given keyโs corresponding entry in the map for in-place manipulation.

See also HashMap::entry.