Module forest_filecoin::cli_shared::cli

Re-exports

pub use self::client::*;

pub use self::config::*;

Modules

client 🔒

config 🔒

Structs

CliOpts

CLI options

CliRpcOpts

CLI RPC options

Enums

ConfigPath

Statics

HELP_MESSAGE

Functions

check_for_unknown_keys

cli_error_and_die

Print an error message and exit the program with an error code Used for handling high level errors such as invalid parameters

find_config_path

find_unknown_keys 🔒