Module forest_filecoin::daemon

Modules

bundle

main

Structs

IPC_PATH 🔒

Functions

__stability 🔒

__static_ref_initialize 🔒

asyncify 🔒

Run the closure on a thread where blocking is allowed

create_password 🔒

Loops until the user provides two matching passwords.

fetch_snapshot_if_required 🔒

If our current chain is below a supported height, we need a snapshot to bring it up to a supported height. If we’ve not been given a snapshot by the user, get one.

get_actual_chain_name 🔒

handle_admin_token 🔒

Generates, prints and optionally writes to a file the administrator JWT token.

input_password_to_load_encrypted_keystore 🔒

Prompts for password, looping until the KeyStore is successfully loaded.

ipc_shmem_conf

load_or_create_keystore 🔒

This may:

propagate_error 🔒

returns the first error with which any of the services end, or never returns at all

start 🔒

Starts daemon process

start_interruptable

unblock_parent_process 🔒