Module forest_filecoin::daemon::db_util

Functions

download_to

filter_lowest_index 🔒

import_chain_as_forest_car

This function validates and stores the CAR binary from from_path(either local path or URL) into the {DB_ROOT}/car_db/ (automatically trans-code into .forest.car.zst format when needed), and returns its final file path and the heaviest tipset.

load_all_forest_cars

move_or_copy_file 🔒

populate_eth_mappings

For the need for Ethereum RPC API, a new column in parity-db has been introduced to handle mapping of:

process_signed_messages 🔒

Filter SignedMessage’s to keep only delegated ones and the most recent ones, then write them to the chain store.

transcode_into_forest_car 🔒