Module forest_filecoin::db

Re-exports

pub use gc::MarkAndSweep;

pub use memory::MemoryDB;

Modules

car

db_engine

db_mode 🔒

gc 🔒

The current implementation of the garbage collector is concurrent mark-and-sweep.

memory 🔒

migration

parity_db

parity_db_config

setting_keys

Traits

DBStatistics

Traits for collecting DB stats

GarbageCollectable

A trait to facilitate mark-and-sweep garbage collection.

SettingsStore

Interface used to store and retrieve settings from the database. To store IPLD blocks, use the BlockStore trait.

SettingsStoreExt

Extension trait for the SettingsStore trait. It is implemented for all types that implement SettingsStore. It provides methods for writing and reading any serializable object from the store.

Functions

truncated_hash 🔒

A function that converts a multihash::MultihashGeneric digest into a u32 representation. We don’t care about collisions here as main use-case is garbage collection.