Module forest_filecoin::db::car

Re-exports

pub use any::AnyCar;

pub use forest::ForestCar;

pub use many::ManyCar;

pub use plain::PlainCar;

Modules

any 🔒

There are three different CAR formats: .car, .car.zst and .forest.car.zst. AnyCar identifies the format by inspecting the CAR header and the first key-value block, and picks the appropriate block store (either super::ForestCar or super::PlainCar).

forest

Forest CAR format

many 🔒

The ManyCar block store is the union of N read-only CAR-backed block stores and a single writable block store. Get requests are forwarded to each store (including the writable store) and the first hit is returned. Write requests are only forwarded to the writable store.

plain

Varint frames

Structs

ZstdFrameCache

Traits

RandomAccessFileReader

Functions

size_of_entry 🔒

Type Definitions

CacheKey

Multiple .forest.car.zst archives may use the same cache, each with a unique cache key.