Module forest_filecoin::db::migration::migration_map

Macros

create_migrations 🔒

A utility macro to make the migrations easier to declare. The usage is: <FROM version> -> <TO version> @ <Migrator object>

Structs

Migration

Statics

MIGRATIONS 🔒

Traits

MigrationOperation 🔒

Migration trait. It is expected that the MigrationOperation::migrate method will pick up the relevant database existing under chain_data_path and create a new migration database in the same directory.

Functions

create_migration_chain 🔒

Creates a migration chain from start to goal. The chain is chosen to be the shortest possible. If there are multiple shortest paths, any of them is chosen. This method will use the pre-defined migrations map.

create_migration_chain_from_migrations 🔒

Same as create_migration_chain, but uses any provided migrations map.

temporary_db_name 🔒

Returns the name of the temporary database that will be created during the migration.

Type Aliases

MigrationsMap 🔒

Migrator 🔒

Migrations map. The key is the starting version and the value is the tuple of the target version and the MigrationOperation implementation.