Module forest_filecoin::db::rolling::gc

The current implementation of the garbage collector is a concurrent, semi-space one.

Design goals

Implement a correct GC algorithm that is simple and efficient for forest scenarios.

GC algorithm

We chose the semi-space GC algorithm for simplicity and sufficiency Besides semi-space, mark-and-sweep was also considered and evaluated. However, it’s not feasible because of the limitations of the underlying DB we use, more specifically, limitations in iterating the DB and retrieving the original key. See https://github.com/paritytech/parity-db/issues/187

GC workflow

Walk back from the current heaviest tipset to the genesis block, collect all the blocks that are reachable from the snapshot

writes blocks that are absent from the current database to it

delete old database(s)

sets current database to a newly created one

Correctness

This algorithm considers all blocks that are visited during the snapshot export task reachable, and ensures they are all transferred and kept in the current DB space. A snapshot can be used to bootstrap a node from scratch thus the algorithm is considered appropriate when the post-GC database contains blocks that are sufficient for exporting a snapshot

Disk usage

During walk_snapshot, data from the old DB is duplicated in the current DB, which uses extra disk space of up to 100% of the snapshot file size

Memory usage

During the data carry-over process, a memory buffer with a fixed capacity is used to speed up the database write operation

Scheduling

GC is triggered automatically when total DB size is greater than 2x of the last reachable data size

GC can be triggered manually by forest-cli db gc command

There’s a global GC lock to ensure at most one GC job is running

Performance

GC performance is typically 1x-1.5x of snapshot export, depending on number of write operations to the current DB space.

Look up performance

DB lookup performance is almost on-par between from single DB and two DBs. Time cost of forest-cli snapshot export --dry-run on DO droplet with 16 GiB RAM is between 9000s to 11000s for both scenarios, no significant performance regression has been observed

Write performance

DB write performance is typically on par with snapshot import. Note that when the current DB space is very large, it tends to trigger DB re-index more frequently, each DB re-index could pause the GC process for a few minutes. The same behavior is observed during snapshot import as well.

Sample mainnet log

2023-03-16T19:50:40.323860Z  INFO crate::db::rolling::gc: Garbage collection started at epoch 2689660
2023-03-16T22:27:36.484245Z  INFO crate::db::rolling::gc: Garbage collection finished at epoch 2689660, took 9416s, reachable data size: 135.71GB
2023-03-16T22:27:38.793717Z  INFO crate::db::rolling::impls: Deleted database under /root/.local/share/forest/mainnet/paritydb/14d0f80992374fb8b20e3b1bd70d5d7b, size: 139.01GB

Structs

DbGarbageCollector

Functions

gc_trigger_factor 🔒