Module forest_filecoin::eth::eip_155_transaction

Structs

EthLegacyEip155TxArgs

Description from Lotus: EthLegacyEip155TxArgs is a legacy Ethereum transaction that uses the EIP-155 chain replay protection mechanism by incorporating the ChainId in the signature. See how the V value in the signature is derived from the ChainId at https://github.com/ethereum/go-ethereum/blob/86a1f0c39494c8f5caddf6bd9fbddd4bdfa944fd/core/types/transaction_signing.go#L424 For EthLegacyEip155TxArgs, the digest that is used to create a signed transaction includes the ChainID but the serialized RLP transaction does not include the ChainID as an explicit field. Instead, the ChainID is included in the V value of the signature as mentioned above.

EthLegacyEip155TxArgsBuilder

Builder for EthLegacyEip155TxArgs.

Enums

EthLegacyEip155TxArgsBuilderError

Error type for EthLegacyEip155TxArgsBuilder

Constants

EIP_155_SIG_PREFIX

Functions

calc_eip155_sig_len 🔒

calc_valid_eip155_sig_len

Returns the valid signature lengths for EIP-155 transactions. The length is based on the chain ID and the V value in the signature.

derive_eip_155_chain_id 🔒

validate_eip155_chain_id 🔒