Module forest_filecoin::fil_cns::validation

Enums

ProofType 🔒

Functionality for verification of seal, winning PoSt and window PoSt proofs. Proof verification will be full validation by default. Verifies winning proof of spacetime. These proofs are generated by the miners that are elected to mine a new block to verify a sector. A failed winning proof leads to a miner being slashed. PoSt proof variants.

Functions

block_sanity_checks 🔒

Checks optional values in header.

block_timestamp_checks 🔒

Check the timestamp corresponds exactly to the number of epochs since the parents.

to_errs 🔒

to_fil_public_replica_infos 🔒

validate_block 🔒

Validates block semantically according to https://github.com/filecoin-project/specs/blob/6ab401c0b92efb6420c6e198ec387cf56dc86057/validation.md Returns all encountered errors, so they can be merged with the common validations performed by the synchronizer.

validate_miner 🔒

validate_ticket_election 🔒

validate_winner_election 🔒

verify_election_post_vrf 🔒

verify_winning_post 🔒

verify_winning_post_proof 🔒