Module forest_filecoin::interpreter::vm

Structs

BlockMessages

Contains all messages to process through the VM as well as miner information for block rewards.

ExecutionContext

MessageCallbackCtx

Enums

CalledAt

VM

Interpreter which handles execution of state transitioning messages and returns receipts from the VM execution.

VMTrace

Tracing a Filecoin VM has a performance penalty. This controls whether a VM should be traced or not when it is created.

Constants

IMPLICIT_MESSAGE_GAS_LIMIT 🔒

Comes from https://github.com/filecoin-project/lotus/blob/v1.23.2/chain/vm/fvm.go#L473

Type Aliases

ForestExecutorV2 🔒

ForestExecutorV3 🔒

ForestExecutorV4 🔒

ForestKernelV2 🔒

ForestKernelV3 🔒

ForestKernelV4 🔒

ForestMachineV2 🔒

ForestMachineV3 🔒

ForestMachineV4 🔒