Module forest_filecoin::ipld::util

Structs

ChainStream

DfsIter

Depth-first-search iterator for ipld leaf nodes.

UnorderedChainStream

Enums

Task 🔒

Constants

BLOCK_CHANNEL_LIMIT 🔒

Functions

recurse_links_hash

Load and hash CIDs and resolve recursively.

should_save_block_to_snapshot 🔒

stream_chain

Stream all blocks that are reachable before the stateroot_limit epoch in a depth-first fashion. After this limit, only block headers are streamed. Any dead links are reported as errors.

stream_graph

traverse_ipld_links_hash 🔒

Traverses all Cid links, hashing and loading all unique values and using the callback function to interact with the data.

unordered_stream_chain

Stream all blocks that are reachable before the stateroot_limit epoch in an unordered fashion. After this limit, only block headers are streamed. Any dead links are reported as errors.

unordered_stream_graph