Module forest_filecoin::key_management::keystore

Modules

json

Structs

EncryptedKeyStore 🔒

Encrypted KeyStore Argon2id hash key derivation XSalsa20Poly1305 authenticated encryption CBOR encoding

KeyInfo

KeyInfo structure, this contains the type of key (stored as a string) and the private key. Note how the private key is stored as a byte vector

KeyStore

KeyStore structure, this contains a set of KeyInfos indexed by address.

PersistentKeyInfo

PersistentKeyStore 🔒

Persistent KeyStore in JSON clear text in KEYSTORE_LOCATION

Enums

EncryptedKeyStoreError

KeyStoreConfig

Constants

CRYPTO_PWHASH_ARGON2ID_MEMLIMIT_INTERACTIVE 🔒

CRYPTO_PWHASH_ARGON2ID_OPSLIMIT_INTERACTIVE 🔒

ENCRYPTED_KEYSTORE_NAME

FOREST_KEYSTORE_PHRASE_ENV

Environmental variable which holds the KeyStore encryption phrase.

KEYSTORE_NAME

NONCE_SIZE 🔒

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

Functions

map_err_to_anyhow 🔒

Type Definitions

SaltByteArray 🔒