Struct forest_filecoin::libp2p::discovery::DiscoveryBehaviour

pub struct DiscoveryBehaviour {
    kademlia: Toggle<Kademlia<MemoryStore>>,
    mdns: Toggle<Mdns>,
    next_kad_random_query: Interval,
    duration_to_next_kad: Duration,
    pending_events: VecDeque<DiscoveryEvent>,
    n_node_connected: u64,
    peers: HashSet<PeerId>,
    peer_addresses: HashMap<PeerId, HashSet<Multiaddr>>,
    target_peer_count: u64,
}

Implementation of NetworkBehaviour that discovers the nodes on the network.

Fields
kademlia: Toggle<Kademlia<MemoryStore>>

Kademlia discovery.

mdns: Toggle<Mdns>

Discovers nodes on the local network.

next_kad_random_query: Interval

Stream that fires when we need to perform the next random Kademlia query.

duration_to_next_kad: Duration

After next_kad_random_query triggers, the next one triggers after this duration.

pending_events: VecDeque<DiscoveryEvent>

Events to return in priority when polled.

n_node_connected: u64

Number of nodes we’re currently connected to.

peers: HashSet<PeerId>

Keeps hash set of peers connected.

peer_addresses: HashMap<PeerId, HashSet<Multiaddr>>

Keeps hash map of peers and their multi-addresses

target_peer_count: u64

Number of connected peers to pause discovery on.

Implementations

impl DiscoveryBehaviour

pub fn peers(&self) -> &HashSet<PeerId>

Returns reference to peer set.

pub fn peer_addresses(&self) -> &HashMap<PeerId, HashSet<Multiaddr>>

Returns a map of peer ids and their multi-addresses

pub fn bootstrap(&mut self) -> Result<QueryId, String>

Bootstrap Kademlia network