Module forest_filecoin::libp2p::service

Modules

metrics 🔒

Structs

Libp2pService

The Libp2pService listens to events from the libp2p swarm.

Topic

A gossipsub topic.

Enums

NetRPCMethods

Network RPC API methods used to gather data from libp2p node.

NetworkEvent

Events emitted by this Service.

NetworkMessage

Messages into the service to handle.

PubsubMessage

Message types that can come over GossipSub

__PrivResult 🔒

Constants

BAN_PEER_DURATION 🔒

BITSWAP_TIMEOUT

PUBSUB_BLOCK_STR

Gossipsub Filecoin blocks topic identifier.

PUBSUB_MSG_STR

Gossipsub Filecoin messages topic identifier.

PUBSUB_TOPICS 🔒

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

Functions

build_transport

Builds the transport stack that libp2p will communicate over. When support of other protocols like udp, quic, http are added, remember to update code comment in Libp2pConfig.

emit_event 🔒

handle_chain_exchange_event 🔒

handle_discovery_event 🔒

handle_forest_behaviour_event 🔒

handle_gossip_event 🔒

handle_hello_event 🔒

handle_network_message 🔒

handle_peer_ops 🔒

handle_ping_event 🔒

Type Definitions

IdentTopic