Module forest_filecoin::libp2p_bitswap

Another libp2p bitswap(SPEC) implementation in Rust.

Features

Compatible with go-bitswap

Optional request manager

Prometheus metrics

Usage

Basic usage of BitswapBehaviour, for writing swarm event flow, sending or receiving a request or a response, checkout tests/go_compat.rs. Note that a request manager is needed for a real-world application.

To use the builtin request manager that is optimized for Filecoin network, a data store that implements BitswapStoreRead and BitswapStoreReadWrite is required. For hooking request manager in swarm event flow, requesting a block via request manager API, checkout tests/request_manager.rs.

Re-exports

pub use metrics::register_metrics;

pub use behaviour::*;

pub use message::*;

pub use store::*;

Modules

behaviour 🔒

bitswap_pb 🔒

internals 🔒

message 🔒

metrics 🔒

request_manager

Request manager implementation that is optimized for filecoin network usage

store 🔒

task

Re-exports API(s) from the chosen task library