Type Alias forest_filecoin::libp2p_bitswap::behaviour::BitswapBehaviourEvent

pub type BitswapBehaviourEvent = Event<Vec<BitswapMessage>, ()>;

libp2p swarm network behavior event of bitswap

Aliased Type

enum BitswapBehaviourEvent {
    Message {
        peer: PeerId,
        message: Message<Vec<BitswapMessage, Global>, (), ()>,
    },
    OutboundFailure {
        peer: PeerId,
        request_id: OutboundRequestId,
        error: OutboundFailure,
    },
    InboundFailure {
        peer: PeerId,
        request_id: InboundRequestId,
        error: InboundFailure,
    },
    ResponseSent {
        peer: PeerId,
        request_id: InboundRequestId,
    },
}

Variants

Message

Fields

peer: PeerId

The peer who sent the message.

message: Message<Vec<BitswapMessage, Global>, (), ()>

The incoming message.

An incoming message (request or response).

OutboundFailure

Fields

peer: PeerId

The peer to whom the request was sent.

request_id: OutboundRequestId

The (local) ID of the failed request.

error: OutboundFailure

The error that occurred.

An outbound request failed.

InboundFailure

Fields

peer: PeerId

The peer from whom the request was received.

request_id: InboundRequestId

The ID of the failed inbound request.

error: InboundFailure

The error that occurred.

An inbound request failed.

ResponseSent

Fields

peer: PeerId

The peer to whom the response was sent.

request_id: InboundRequestId

The ID of the inbound request whose response was sent.

A response to an inbound request has been sent.

When this event is received, the response has been flushed on the underlying transport connection.