Trait forest_filecoin::libp2p_bitswap::BitswapStoreRead

pub trait BitswapStoreRead {
    // Required methods
    fn contains(&self, cid: &Cid) -> Result<bool>;
    fn get(&self, cid: &Cid) -> Result<Option<Vec<u8>>>;
}

Trait implemented by a block store for reading.

Required Methods

fn contains(&self, cid: &Cid) -> Result<bool>

A have query needs to know if the block store contains the block.

fn get(&self, cid: &Cid) -> Result<Option<Vec<u8>>>

A block query needs to retrieve the block from the store.