Trait forest_filecoin::lotus_json::HasLotusJson

pub trait HasLotusJson: Sized + Into<Self::LotusJson> {
    type LotusJson: Into<Self> + Serialize + DeserializeOwned;

    // Required method
    fn snapshots() -> Vec<(Value, Self)>;
}

Required Associated Types

type LotusJson: Into<Self> + Serialize + DeserializeOwned

The struct representing JSON. You should #[derive(Deserialize, Serialize)] on it.

Required Methods

fn snapshots() -> Vec<(Value, Self)>

To ensure code quality, conversion to/from lotus JSON MUST be tested. Provide snapshots of the JSON, and the domain type it should serialize to.

Serialization and de-serialization of the domain type should match the snapshot.

If using decl_and_test, this test is automatically run for you, but if the test is out-of-module, you must call assert_all_snapshots manually.