Struct forest_filecoin::message_pool::msg_chain::Chains

pub(super) struct Chains {
    pub map: SlotMap<NodeKey, MsgChainNode>,
    pub key_vec: Vec<NodeKey>,
}

Chains is an abstraction of a list of message chain nodes. It wraps a SlotMap instance. key_vec is an additional requirement in order to satisfy optimal msg selection use cases, such as iteration in insertion order. The SlotMap serves as a lookup table for nodes to get around the borrow checker rules. Each MsgChainNode contains only pointers as NodeKey to the entries in the map With this design, we get around the borrow checker rule issues when implementing the optimal selection algorithm.

Fields
map: SlotMap<NodeKey, MsgChainNode>key_vec: Vec<NodeKey>
Implementations

impl Chains

pub(super) fn new() -> Self

pub(super) fn push_with( &mut self, cur_chain: MsgChainNode, node_vec: &mut Vec<NodeKey> )

Pushes a msg chain node into slot map and places the key in the node_vec passed as parameter.

pub(super) fn sort(&mut self, rev: bool)

Sorts the chains with compare method. If rev is true, sorts in descending order.

pub(super) fn get_mut(&mut self, k: NodeKey) -> Option<&mut MsgChainNode>

Retrieves the msg chain node by the given NodeKey

pub(super) fn get_mut_at(&mut self, i: usize) -> Option<&mut MsgChainNode>

Retrieves the msg chain node at the given index

pub(super) fn get_from(&self, i: usize, vec: &[NodeKey]) -> &MsgChainNode

pub(super) fn get_mut_from( &mut self, i: usize, vec: &[NodeKey] ) -> &mut MsgChainNode

pub(super) fn get_key_at(&self, i: usize) -> Option<NodeKey>

pub(super) fn get_at(&mut self, i: usize) -> Option<&MsgChainNode>

Retrieves the msg chain node at the given index

pub(super) fn len(&self) -> usize

Retrieves the amount of items.

pub(super) fn is_empty(&self) -> bool

Returns true is the chain is empty and otherwise. We check the map as the source of truth as key_vec can be extended time to time.

pub(super) fn trim_msgs_at( &mut self, idx: usize, gas_limit: u64, base_fee: &TokenAmount )

Removes messages from the given index and resets effective perfs

pub(super) fn invalidate(&mut self, key: Option<NodeKey>)

pub(super) fn drop_invalid(&mut self, key_vec: &mut Vec<NodeKey>)

Drops nodes which are no longer valid after the merge step