Function forest_filecoin::message_pool::msgpool::head_change

pub async fn head_change<T>(
    api: &T,
    bls_sig_cache: &Mutex<LruCache<Cid, Signature>>,
    repub_trigger: Arc<Sender<()>>,
    republished: &RwLock<HashSet<Cid>>,
    pending: &RwLock<HashMap<Address, MsgSet>>,
    cur_tipset: &Mutex<Arc<Tipset>>,
    revert: Vec<Tipset>,
    apply: Vec<Tipset>
) -> Result<(), Error>
where
    T: Provider + 'static,

This function will revert and/or apply tipsets to the message pool. This function should be called every time that there is a head change in the message pool.