Module forest_filecoin::networks::butterflynet

Constants

BREEZE_GAS_TAMPING_DURATION

ETH_CHAIN_ID

GENESIS_NETWORK_VERSION

MINIMUM_CONSENSUS_POWER 🔒

MINIMUM_VERIED_ALLOCATION 🔒

PRE_COMMIT_CHALLENGE_DELAY 🔒

Statics

DEFAULT_BOOTSTRAP

Default bootstrap peer ids.

DRAND_SCHEDULE 🔒

GENESIS_CID

Genesis CID

GENESIS_URL 🔒

Compressed genesis file. It is compressed with zstd and cuts the download size by 80% (from 10 MB to 2 MB).

GENESIS_URL_ALT 🔒

Alternative URL for the genesis file. This is hosted on the lotus repository and is not compressed. The genesis file does not live on the master branch, currently on a draft PR. <https://github.com/filecoin-project/lotus/pull/11458>

HEIGHT_INFOS

Height epochs.

Functions

fetch_genesis

Fetches the genesis CAR from the local database or downloads it if it does not exist. The result bytes may be compressed.