Module forest_filecoin::rpc

Re-exports

pub use client::Client;

pub use error::ServerError;

pub use reflect::ApiVersion;

pub use reflect::RpcMethod;

pub use reflect::RpcMethodExt;

pub use request::Request;

pub use crate::rpc::channel::CANCEL_METHOD_NAME;

pub use methods::*;

Modules

auth_layer 🔒

channel 🔒

Subscription related types and traits for server implementations.

client 🔒

Design Goals

error 🔒

methods 🔒

All the methods live in their own folder

prelude

All handler definitions.

reflect 🔒

Forest wishes to provide OpenRPC definitions for Filecoin APIs. To do this, it needs:

request 🔒

types

Types that are shared between APIs.

Macros

for_each_method 🔒

The macro callback will be passed in each type that implements RpcMethod.

Structs

PerConnection 🔒

RPCState

This is where you store persistent data, or at least access to stateful data.

Enums

ClientError

Protocol or transport-specific error Error type.

Constants

DEFAULT_PORT

MAX_REQUEST_BODY_SIZE 🔒

MAX_RESPONSE_BODY_SIZE 🔒

Statics

DEFAULT_REQUEST_TIMEOUT 🔒

Request timeout read from environment variables

Functions

create_module 🔒

openrpc

start_rpc