Module forest_filecoin::rpc::methods::eth

Modules

types

Structs

Block

BlockHash

BlockNumber

Bloom

EthBigInt

EthSyncingResult

Hash

Int64

Nonce

Tx

TxArgs 🔒

Uint64

Enums

BlockNumberOrHash

EAMMethod 🔒

EVMMethod 🔒

EthAccounts

EthBlockNumber

EthChainId

EthGasPrice

EthGetBalance

EthGetBlockByHash

EthGetBlockByNumber

EthGetBlockTransactionCountByHash

EthGetBlockTransactionCountByNumber

EthGetCode

EthGetStorageAt

EthSyncing

EthSyncingResultLotusJson

Predefined

Transactions

Web3ClientVersion

Constants

ADDRESS_LENGTH 🔒

Ethereum address size in bytes.

BLOOM_SIZE 🔒

Ethereum Bloom filter size in bits. Bloom filter is used in Ethereum to minimize the number of block queries.

BLOOM_SIZE_IN_BYTES 🔒

Ethereum Bloom filter size in bytes.

EIP_1559_TX_TYPE 🔒

Ethereum Improvement Proposals 1559 transaction type. This EIP changed Ethereum’s fee market mechanism. Transaction type can have 3 distinct values:

EMPTY_ROOT 🔒

Keccak-256 of the RLP of null.

EMPTY_UNCLES 🔒

Keccak-256 of an RLP of an empty array. In Filecoin, we don’t have the concept of uncle blocks but rather use tipsets to reward miners who craft blocks.

EVM_WORD_LENGTH 🔒

Ethereum Virtual Machine word size in bytes.

FULL_BLOOM 🔒

Ethereum Bloom filter with all bits set to 1.

MASKED_ID_PREFIX 🔒

REVERTED_ETH_ADDRESS 🔒

The address used in messages to actors that have since been deleted.

Functions

block_from_filecoin_tipset

count_messages_in_tipset 🔒

decode_payload 🔒

Decodes the payload using the given codec.

encode_as_abi_helper 🔒

Format two numbers followed by an arbitrary byte array as solidity ABI.

encode_filecoin_params_as_abi 🔒

See https://docs.soliditylang.org/en/latest/abi-spec.html#function-selector-and-argument-encoding for ABI specification

encode_filecoin_returns_as_abi 🔒

eth_tx_args_from_unsigned_eth_message 🔒

eth_tx_from_native_message 🔒

Convert a native message to an eth transaction.

eth_tx_from_signed_eth_message

eth_tx_from_signed_eth_message does NOT populate:

execute_tipset 🔒

format_address 🔒

format_bigint 🔒

format_u64 🔒

get_tipset_from_hash 🔒

is_eth_address 🔒

lookup_eth_address 🔒

new_eth_tx_from_signed_message

recover_sig 🔒

round_up_word 🔒

Round to the next multiple of EVM word length.

tipset_by_block_number_or_hash 🔒

Type Aliases

GasPriceResult 🔒