Module forest_filecoin::rpc::methods::f3

This module contains F3(fast finality) related V1 RPC methods as well as some internal RPC methods(F3.*) that power the go-f3 node in sidecar mode.

Modules

types 🔒

util 🔒

Enums

F3GetCertificate

returns a finality certificate at given instance number

F3GetECPowerTable

F3GetF3PowerTable

F3GetLatestCertificate

returns the latest finality certificate

F3Participate

F3Participate should be called by a storage provider to participate in signing F3 consensus. Calling this API gives the node a lease to sign in F3 on behalf of given SP. The lease should be active only on one node. The lease will expire at the newLeaseExpiration. To continue participating in F3 with the given node, call F3Participate again before the newLeaseExpiration time. newLeaseExpiration cannot be further than 5 minutes in the future. It is recommended to call F3Participate every 60 seconds with newLeaseExpiration set 2min into the future. The oldLeaseExpiration has to be set to newLeaseExpiration of the last successful call. For the first call to F3Participate, set the oldLeaseExpiration to zero value/time in the past. F3Participate will return true if the lease was accepted. The minerID has to be the ID address of the miner.

GetHead

GetParent

GetParticipatingMinerIDs

GetPowerTable

GetTipset

GetTipsetByEpoch

ProtectPeer

SignMessage

Statics

F3_LEASE_MANAGER 🔒

Functions

get_f3_rpc_endpoint

get_rpc_http_client 🔒