Module forest_filecoin::rpc::methods::state

Macros

for_each_method 🔒

Enums

StateGetBeaconEntry

Constants

MINER_GET_BASE_INFO

MSIG_GET_AVAILABLE_BALANCE

MSIG_GET_PENDING

STATE_ACCOUNT_KEY

STATE_CALL

STATE_CIRCULATING_SUPPLY

STATE_DEAL_PROVIDER_COLLATERAL_BOUNDS

STATE_DECODE_PARAMS

STATE_FETCH_ROOT

STATE_GET_ACTOR

STATE_GET_RANDOMNESS_FROM_BEACON

STATE_GET_RANDOMNESS_FROM_TICKETS

STATE_GET_RECEIPT

STATE_LIST_MESSAGES

STATE_LIST_MINERS

STATE_LOOKUP_ID

STATE_MARKET_BALANCE

STATE_MARKET_DEALS

STATE_MARKET_STORAGE_DEAL

STATE_MINER_ACTIVE_SECTORS

STATE_MINER_AVAILABLE_BALANCE

STATE_MINER_DEADLINES

STATE_MINER_FAULTS

STATE_MINER_INFO

STATE_MINER_PARTITIONS

STATE_MINER_POWER

STATE_MINER_PROVING_DEADLINE

STATE_MINER_RECOVERIES

STATE_MINER_SECTORS

STATE_MINER_SECTOR_COUNT

STATE_NETWORK_NAME

STATE_NETWORK_VERSION

STATE_READ_STATE

STATE_REPLAY

STATE_SEARCH_MSG

STATE_SEARCH_MSG_LIMITED

STATE_SECTOR_GET_INFO

STATE_VERIFIED_CLIENT_STATUS

STATE_VM_CIRCULATING_SUPPLY_INTERNAL

STATE_WAIT_MSG

Functions

lock_pop 🔒

miner_get_base_info

msig_get_available_balance

msig_get_pending

state_account_key

gets the public key address of the given ID address See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateAccountKey

state_call

runs the given message and returns its result without any persisted changes.

state_circulating_supply

state_deal_provider_collateral_bounds

state_fetch_root

Traverse an IPLD directed acyclic graph and use libp2p-bitswap to request any missing nodes. This function has two primary uses: (1) Downloading specific state-roots when Forest deviates from the mainline blockchain, (2) fetching historical state-trees to verify past versions of the consensus rules.

state_get_actor 🔒

state_get_network_version

state_get_randomness_from_beacon

Get randomness from beacon

state_get_randomness_from_tickets

Get randomness from tickets

state_get_receipt

returns the message receipt for the given message

state_list_messages 🔒

Looks back and returns all messages with a matching to or from address, stopping at the given height.

state_list_miners

state_lookup_id

retrieves the ID address of the given address See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateLookupID

state_market_balance

looks up the Escrow and Locked balances of the given address in the Storage Market

state_market_deals

state_market_storage_deal

state_miner_active_sectors

state_miner_available_balance

state_miner_deadlines

state_miner_faults

looks up the miner power of the given address.

state_miner_info

looks up the miner info of the given address.

state_miner_partitions

state_miner_power

looks up the miner power of the given address.

state_miner_proving_deadline

state_miner_recoveries

state_miner_sector_count

state_miner_sectors

state_network_name

gets network name from state manager

state_read_state

Get read state

state_replay

returns the result of executing the indicated message, assuming it was executed in the indicated tipset.

state_search_msg

Searches for a message in the chain, and returns its receipt and the tipset where it was executed. See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateSearchMsg

state_search_msg_limited

Looks back up to limit epochs in the chain for a message, and returns its receipt and the tipset where it was executed. See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateSearchMsgLimited

state_sector_get_info

Get state sector info using sector no

state_verified_client_status 🔒

state_vm_circulating_supply_internal 🔒

state_wait_msg

looks back in the chain for a message. If not found, it blocks until the message arrives on chain, and gets to the indicated confidence depth.

Type Aliases

RandomnessParams 🔒