Module forest_filecoin::rpc::reflect

Forest wishes to provide OpenRPC definitions for Filecoin APIs. To do this, it needs:

JSON Schema definitions for all the argument and return types.

The number of arguments (arity) and names of those arguments for each RPC method.

As a secondary objective, we wish to provide an RPC client for our CLI, and internal tests against Lotus.

The RpcMethod trait encapsulates all the above at a single site.

schemars::JsonSchema provides schema definitions,

RpcMethod defining arity and actually dispatching the function calls.

Modules

jsonrpc_types

A transcription of types from the JSON-RPC 2.0 Specification.

parser 🔒

util 🔒

Macros

do_impls 🔒

Enums

ApiPath

ApiPaths

Which paths should this method be exposed on?

ConcreteCallingConvention

openrpc_types::ParamStructure describes accepted param format. This is an actual param format, used to decide how to construct arguments.

Permission

The permission required to call an RPC method.

Traits

Params

A tuple of ARITY arguments.

RpcMethod

A definition of an RPC method handler which:

RpcMethodExt

Utility methods, defined as an extension trait to avoid having to specify ARITY in user code.

Functions

unexpected 🔒

Type Aliases

Ctx

Type to be used by RpcMethod::handle.