Module forest_filecoin::rpc::reflect::openrpc_types

A transcription of types from the OpenRPC Specification.

When quoted, the specification will appear as block-quoted text, like so.

Macros

validated_vec 🔒

parent acts as a Vec<$child>, with some internal invariant guarded by $parent::new().

Structs

Components

ContentDescriptor

Method

Methods

MethodsError

OpenRPC

ParamListError

Params

REQUIRED. A list of parameters that are applicable for this method. The list MUST NOT include duplicated parameters and therefore require name to be unique. The list can use the Reference Object to link to parameters that are defined by the Content Descriptor Object. All optional params (content descriptor objects with “required”: false) MUST be positioned after all required params in the list.

Enums

ParamStructure

The expected format of the parameters. As per the JSON-RPC 2.0 specification, the params of a JSON-RPC request object may be an array, object, or either (represented as by-position, by-name, and either respectively). When a method has a paramStructure value of by-name, callers of the method MUST send a JSON-RPC request object whose params field is an object. Further, the key names of the params object MUST be the same as the contentDescriptor.names for the given method. Defaults to “either”.