Trait forest_filecoin::rpc::reflect::RpcMethodExt

pub trait RpcMethodExt<const ARITY: usize>: RpcMethod<ARITY> {
    // Provided methods
    fn build_params(
        params: Self::Params,
        calling_convention: ConcreteCallingConvention
    ) -> Result<RequestParameters, Error> { ... }
    fn openrpc<'de>(
        gen: &mut SchemaGenerator,
        calling_convention: ParamStructure
    ) -> Result<Method, ParamListError>
       where <Self::Ok as HasLotusJson>::LotusJson: JsonSchema + Deserialize<'de> { ... }
    fn register_raw(
        module: &mut RpcModule<RPCState<impl Blockstore + Send + Sync + 'static>>,
        calling_convention: ParamStructure
    ) -> Result<&mut MethodCallback, RegisterMethodError>
       where <Self::Ok as HasLotusJson>::LotusJson: Clone + 'static { ... }
    fn register<'de>(
        module: &mut SelfDescribingRpcModule<RPCState<impl Blockstore + Send + Sync + 'static>>
    )
       where <Self::Ok as HasLotusJson>::LotusJson: Clone + JsonSchema + Deserialize<'de> + 'static { ... }
    fn call_module(
        module: &RpcModule<Arc<RPCState<impl Blockstore + Send + Sync + 'static>>>,
        params: Self::Params
    ) -> impl Future<Output = Result<Self::Ok, MethodsError>> + Send
       where Self::Ok: DeserializeOwned + Clone + Send { ... }
    fn request(params: Self::Params) -> Result<RpcRequest<Self::Ok>, Error> { ... }
    fn call_raw(
        client: &Client,
        params: Self::Params
    ) -> impl Future<Output = Result<<Self::Ok as HasLotusJson>::LotusJson, ClientError>> { ... }
    fn call(
        client: &Client,
        params: Self::Params
    ) -> impl Future<Output = Result<Self::Ok, ClientError>> { ... }
}

Utility methods, defined as an extension trait to avoid having to specify ARITY in user code.

Provided Methods

fn build_params( params: Self::Params, calling_convention: ConcreteCallingConvention ) -> Result<RequestParameters, Error>

Convert from typed handler parameters to un-typed JSON-RPC parameters.

Exposes errors from Params::unparse

fn openrpc<'de>( gen: &mut SchemaGenerator, calling_convention: ParamStructure ) -> Result<Method, ParamListError>
where <Self::Ok as HasLotusJson>::LotusJson: JsonSchema + Deserialize<'de>,

Generate a full OpenRPC method definition for this endpoint.

fn register_raw( module: &mut RpcModule<RPCState<impl Blockstore + Send + Sync + 'static>>, calling_convention: ParamStructure ) -> Result<&mut MethodCallback, RegisterMethodError>
where <Self::Ok as HasLotusJson>::LotusJson: Clone + 'static,

Register this method with an RpcModule.

fn register<'de>( module: &mut SelfDescribingRpcModule<RPCState<impl Blockstore + Send + Sync + 'static>> )
where <Self::Ok as HasLotusJson>::LotusJson: Clone + JsonSchema + Deserialize<'de> + 'static,

Register this method and generate a schema entry for it in a SelfDescribingRpcModule.

fn call_module( module: &RpcModule<Arc<RPCState<impl Blockstore + Send + Sync + 'static>>>, params: Self::Params ) -> impl Future<Output = Result<Self::Ok, MethodsError>> + Send
where Self::Ok: DeserializeOwned + Clone + Send,

Call this method on an RpcModule.

fn request(params: Self::Params) -> Result<RpcRequest<Self::Ok>, Error>

Returns Err if any of the parameters fail to serialize.

fn call_raw( client: &Client, params: Self::Params ) -> impl Future<Output = Result<<Self::Ok as HasLotusJson>::LotusJson, ClientError>>

fn call( client: &Client, params: Self::Params ) -> impl Future<Output = Result<Self::Ok, ClientError>>

Object Safety

This trait is not object safe.