Trait forest_filecoin::rpc::reflect::Wrap

pub trait Wrap<const ARITY: usize, Ctx, Args, R> {
    type Future: Future<Output = Result<Value, JsonRpcError>> + Send;

    // Required method
    fn wrap(
        self,
        param_names: [&'static str; ARITY],
        calling_convention: ParamStructure
    ) -> impl Clone + Send + Sync + 'static + Fn(Params<'static>, Arc<Ctx>) -> Self::Future;
}

Wrap a bare function with our argument parsing logic. Turns any async fn foo(ctx, arg0...) into a function that can be passed to jsonrpsee::server::RpcModule::register_async_method.

This trait is NOT designed to implemented on anything except bare handler functions, as below.

Required Associated Types

type Future: Future<Output = Result<Value, JsonRpcError>> + Send

Required Methods

fn wrap( self, param_names: [&'static str; ARITY], calling_convention: ParamStructure ) -> impl Clone + Send + Sync + 'static + Fn(Params<'static>, Arc<Ctx>) -> Self::Future

Object Safety

This trait is not object safe.