Module forest_filecoin::rpc::state_api

Functions

lock_pop ğŸ”’

state_call ğŸ”’

runs the given message and returns its result without any persisted changes.

state_fetch_root ğŸ”’

Traverse an IPLD directed acyclic graph and use libp2p-bitswap to request any missing nodes. This function has two primary uses: (1) Downloading specific state-roots when Forest deviates from the mainline blockchain, (2) fetching historical state-trees to verify past versions of the consensus rules.

state_get_actor ğŸ”’

state_get_network_version ğŸ”’

state_get_randomness_from_beacon ğŸ”’

Get randomness from beacon

state_get_receipt ğŸ”’

returns the message receipt for the given message

state_market_balance ğŸ”’

looks up the Escrow and Locked balances of the given address in the Storage Market

state_market_deals ğŸ”’

state_miner_power ğŸ”’

looks up the miner power of the given address.

state_network_name ğŸ”’

gets network name from state manager

state_replay ğŸ”’

returns the result of executing the indicated message, assuming it was executed in the indicated tipset.

state_wait_msg ğŸ”’

looks back in the chain for a message. If not found, it blocks until the message arrives on chain, and gets to the indicated confidence depth.

Type Aliases

RandomnessParams ğŸ”’