Module forest_filecoin::rpc::wallet_api

Functions

wallet_balance

Return the balance from StateManager for a given Address

wallet_default_address

Get the default Address for the Wallet

wallet_delete

Deletes a wallet given its address.

wallet_export

Export KeyInfo from the Wallet given its address

wallet_has

Return whether or not a Key is in the Wallet

wallet_import

Import KeyInfo to the Wallet, return the Address that corresponds to it

wallet_list

List all Addresses in the Wallet

wallet_new

Generate a new Address that is stored in the Wallet

wallet_set_default

Set the default Address for the Wallet

wallet_sign

Sign a vector of bytes

wallet_validate_address 🔒

Validates whether a given string can be decoded as a well-formed address

wallet_verify

Verify a Signature, true if verified, false otherwise