Module forest_filecoin::rpc_api

In general, forest wants to support the same RPC messages as lotus (go implementation of Filecoin).

Follow the pattern set below, and don’t forget to add an entry to the ACCESS_MAP with the relevant permissions (consult the go implementation, looking for a comment like // perm: admin)

Future work:

Have an RpcEndpoint trait.

Modules

auth_api

JSON-RPC API definitions Authorization API

beacon_api

Beacon API

chain_api

Chain API

common_api

Common API

data_types

gas_api

Gas API

mpool_api

Message Pool API

net_api

Net API

node_api

Node API

state_api

State API

sync_api

Sync API

wallet_api

Wallet API

Enums

Access

Access levels to be checked against JWT claims

Statics

ACCESS_MAP

Access mapping between method names and access levels Checked against JWT claims on every request

Functions

check_access

Checks an access enumeration against provided JWT claims