Module forest_filecoin::rpc_api::data_types

Structs

APIVersion

Represents the current version of the API.

ActorStateJson

State of all actor implementations.

AddrInfo

AddressOrEmpty

This wrapper is needed because of a bug in Lotus. See: https://github.com/filecoin-project/lotus/issues/11461.

ApiActorState

ApiDeadline

ApiHeadChange

ApiInvocResult

ApiMessage

ApiMessageLotusJson

ApiReceipt

ApiState 🔒

ApiTipsetKey

BeneficiaryTermLotusJson

BlockMessages

CirculatingSupply

DiscoverDocs

DiscoverInfo

DiscoverMethod

DiscoverResult

ExecutionTrace

GasTrace

MarketDeal

MessageFilter

MessageGasCost

MessageLookup

MessageSendSpec

MessageTrace

MinerInfoLotusJson

MinerPowerLotusJson

MinerSectors

MiningBaseInfo

Multiaddr

Representation of a Multiaddr.

PeerID

PendingBeneficiaryChangeLotusJson

RPCState

This is where you store persistent data, or at least access to stateful data.

RPCSyncState

ReturnTrace

SectorOnChainInfo

Transaction

Version

Integer based value on version information. Highest order bits for Major, Mid order for Minor and lowest for Patch.

Constants

EMPTY_ADDRESS_VALUE 🔒

Type Aliases

Data