Module forest_filecoin::shim::actors::version

Macros

impl_actor_cids 🔒

impl_actor_cids_type_actor 🔒

Statics

ACCOUNT_ACTOR_CIDS 🔒

CRON_ACTOR_CIDS 🔒

DATACAP_ACTOR_CIDS 🔒

EAM_ACTOR_CIDS 🔒

ETHACCOUNT_ACTOR_CIDS 🔒

EVM_ACTOR_CIDS 🔒

INIT_ACTOR_CIDS 🔒

MARKET_ACTOR_CIDS 🔒

MINER_ACTOR_CIDS 🔒

MULTISIG_ACTOR_CIDS 🔒

PAYMENTCHANNEL_ACTOR_CIDS 🔒

PLACEHOLDER_ACTOR_CIDS 🔒

POWER_ACTOR_CIDS 🔒

REWARD_ACTOR_CIDS 🔒

SYSTEM_ACTOR_CIDS 🔒

VERIFREG_ACTOR_CIDS 🔒

Functions

is_account_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_account_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_cron_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_cron_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_datacap_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_datacap_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_eam_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_eam_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_ethaccount_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_ethaccount_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_evm_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_evm_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_init_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_init_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_market_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_market_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_miner_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_miner_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_multisig_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_multisig_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_paymentchannel_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_paymentchannel_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_placeholder_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_placeholder_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_power_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_power_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_reward_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_reward_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_system_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_system_cid_version

Checks if the provided actor code CID and version are valid for the given actor.

is_verifreg_actor

Checks if the provided actor code CID is valid for the given actor (any version).

is_verifreg_cid_version

Checks if the provided actor code CID and version are valid for the given actor.