Module forest_filecoin::shim::address

Structs

Address

A Filecoin address is an identifier that refers to an actor in the Filecoin state. All actors (miner actors, the storage market actor, account actors) have an address. This address encodes information about the network to which an actor belongs, the specific type of address encoding, the address payload itself, and a checksum. The goal of this format is to provide a robust address format that is both easy to use and resistant to errors.

CurrentNetwork

For user safety, Filecoin has different addresses for its mainnet and test networks: Mainnet and testnet addresses are prefixed with f and t, respectively.

NetworkGuard 🔒

StrictAddress

A Filecoin address is an identifier that refers to an actor in the Filecoin state. All actors (miner actors, the storage market actor, account actors) have an address. This address encodes information about the network to which an actor belongs, the specific type of address encoding, the address payload itself, and a checksum. The goal of this format is to provide a robust address format that is both easy to use and resistant to errors.

Enums

Error

Address error

Network

Network defines the preconfigured networks to use with address encoding

Payload

Payload is the data of the Address. Variants are the supported Address protocols.

Protocol

Protocol defines the addressing protocol used to derive data to an address

Constants

ADDRESS_ENCODER 🔒

defines the encoder for base32 encoding with the provided string with no padding

LOCAL_NETWORK 🔒

PAYLOAD_HASH_LEN

Hash length of payload for Secp and Actor addresses.

Statics

GLOBAL_NETWORK 🔒

ZERO_ADDRESS

Zero address used to avoid allowing it to be used for verification. This is intentionally disallowed because it is an edge case with Filecoin’s BLS signature verification.