Module forest_filecoin::shim::fvm_latest

Modules

call_manager

engine

executor

externs

This module contains the logic to invoke the node by traversing Boundary A.

gas

kernel

machine

state_tree

syscalls

trace

Macros

syscall_error

Convenience macro for generating Actor Errors

total_enum_map

Create a mapping from enum items to values in a way that guarantees at compile time that we did not miss any member, in any of the prices, even if the enum gets a new member later.

Structs

DefaultKernel

The “default” Kernel implementation.

EMPTY_ARR_CID

Cid of the empty array Cbor bytes (EMPTY_ARR_BYTES).

Traits

Kernel

The “kernel” implements the FVM interface as presented to the actors. It: