Struct forest_filecoin::shim::fvm_latest::engine::Engine

pub struct Engine {
    id: u64,
    inner: Arc<EngineInner, Global>,
}

An Engine represents a single, caching wasm engine. It should not be shared between concurrent call stacks.

The Engine will be returned to the EnginePool on drop.

Fields
id: u64inner: Arc<EngineInner, Global>
Implementations

impl Engine

pub fn prepare_actor_code<BS>( &self, code_cid: &Cid<64>, blockstore: BS ) -> Result<usize, Error>where BS: Blockstore,

Loads an actor’s Wasm code from the blockstore by CID, and prepares it for execution by instantiating and caching the Wasm module. This method errors if the code CID is not found in the store.

Return the original byte code size.

pub fn preload<'a, BS, I>( &self, blockstore: BS, cids: I ) -> Result<usize, Error>where BS: Blockstore, I: IntoIterator<Item = &'a Cid<64>>,

Instantiates and caches the Wasm modules for the bytecodes addressed by the supplied CIDs. Only uncached entries are actually fetched and instantiated. Blockstore failures and entry inexistence shortcircuit make this method return an Err immediately.

Returns the total original byte size of the modules

pub fn prepare_wasm_bytecode( &self, k: &Cid<64>, wasm: &[u8] ) -> Result<usize, Error>

Loads some Wasm code into the engine and prepares it for execution.

pub unsafe fn load_compiled( &self, k: &Cid<64>, compiled: &[u8] ) -> Result<Module, Error>

Load compiled wasm code into the engine.

Safety

See [wasmtime::Module::deserialize] for safety information.

pub fn get_module( &self, blockstore: &impl Blockstore, k: &Cid<64> ) -> Result<Option<Module>, Error>

Lookup a loaded wasmtime module.

pub fn instantiate<K>( &self, store: &mut Store<InvocationData<K>>, k: &Cid<64> ) -> Result<Option<Instance>, Abort>where K: Kernel,

Lookup and instantiate a loaded wasmtime module with the given store. This will cache the linker, syscalls, etc.

This returns an Abort as it may need to execute initialization code, charge gas, etc.

pub fn new_store<K>(&self, kernel: K) -> Store<InvocationData<K>>where K: Kernel,

Construct a new wasmtime “store” from the given kernel.