Module forest_filecoin::shim::fvm_latest::executor

Structs

ApplyRet

Apply message return data.

DefaultExecutor

The default Executor.

ThreadedExecutor

An executor that executes messages on a separate thread with a 64MiB stack. If you can guarantee at least 64MiB of stack space, you don’t need this executor.

Enums

ApplyFailure

A description of some failure encountered when applying a message.

ApplyKind

The kind of message being applied:

Traits

Executor

An executor executes messages on the underlying machine/kernel. It’s responsible for: