Trait forest_filecoin::shim::fvm_latest::executor::Executor

pub trait Executor {
    type Kernel: Kernel;

    // Required methods
    fn execute_message(
        &mut self,
        msg: Message,
        apply_kind: ApplyKind,
        raw_length: usize
    ) -> Result<ApplyRet, Error>;
    fn flush(&mut self) -> Result<Cid<64>, Error>;
}

An executor executes messages on the underlying machine/kernel. It’s responsible for:

Validating messages (nonce, sender, etc).

Creating message receipts.

Charging message inclusion gas, overestimation gas, miner tip, etc.

Required Associated Types

type Kernel: Kernel

The Kernel on which messages will be applied. We specify a Kernel here, not a Machine, because the Kernel implies the Machine.

Required Methods

fn execute_message( &mut self, msg: Message, apply_kind: ApplyKind, raw_length: usize ) -> Result<ApplyRet, Error>

This is the entrypoint to execute a message.

NOTE: The “raw length” is the length of the message as it appears on-chain and is used to charge message inclusion gas.

fn flush(&mut self) -> Result<Cid<64>, Error>

Flushes the state-tree, returning the new root CID.