Module forest_filecoin::shim::fvm_latest::kernel

Modules

default

Structs

Block

BlockRegistry

A registry of open blocks (per-kernel). Think “file descriptor” table. At the moment, there’s no way to close/remove a block from this table.

BlockStat

SendResult

SyscallError

Represents an error from a syscall. It can optionally contain a syscall-advised exit code for the kind of error that was raised. We may want to add an optional source error here.

Enums

ExecutionError

SupportedHashes

Codes and hashers supported by FVM. You can use this hash directly inside of your actor, but it will very likely be more performant with the hash syscall

Traits

ActorOps

Actors operations whose scope of action is actors other than the calling actor. The calling actor’s state may be consulted to resolve some.

CircSupplyOps

Operations to query the circulating supply.

ClassifyResult

Context

The FVM’s equivalent of anyhow::Context. This is intentionally only implemented on ExecutionError and Result<T, ExecutionError> so anyhow::Context can be imported at the same time.

CryptoOps

Cryptographic primitives provided by the kernel.

DebugOps

Debugging APIs.

EventOps

Eventing APIs.

GasOps

Operations for explicit gas charging.

IpldBlockOps

The IPLD subset of the kernel.

Kernel

The “kernel” implements the FVM interface as presented to the actors. It:

LimiterOps

Track and limit memory expansion.

MessageOps

Accessors to query attributes of the incoming message.

NetworkOps

Network-related operations.

RandomnessOps

Randomness queries.

SelfOps

Actor state access and manipulation. Depends on BlockOps to read and write blocks in the state tree.

Type Aliases

BlockId

Blocks in the block registry are addressed by an ordinal, starting from 1 (FIRST_ID). The zero value is reserved to mean “no data”, such as when actor invocations receive or return no data.

Result

Execution result.