Module forest_filecoin::shim::fvm_latest::kernel::prelude

Import this module (with a glob) if you’re implementing a kernel, especially if you want to use ambassador to delegate the implementation.

Macros

ambassador_impl_ActorOps

A macro to be used by ambassador::Delegate to delegate ActorOps

ambassador_impl_CryptoOps

A macro to be used by ambassador::Delegate to delegate CryptoOps

ambassador_impl_DebugOps

A macro to be used by ambassador::Delegate to delegate DebugOps

ambassador_impl_EventOps

A macro to be used by ambassador::Delegate to delegate EventOps

ambassador_impl_IpldBlockOps

A macro to be used by ambassador::Delegate to delegate IpldBlockOps

ambassador_impl_MessageOps

A macro to be used by ambassador::Delegate to delegate MessageOps

ambassador_impl_NetworkOps

A macro to be used by ambassador::Delegate to delegate NetworkOps

ambassador_impl_RandomnessOps

A macro to be used by ambassador::Delegate to delegate RandomnessOps

ambassador_impl_SelfOps

A macro to be used by ambassador::Delegate to delegate SelfOps

ambassador_impl_SendOps

A macro to be used by ambassador::Delegate to delegate SendOps

ambassador_impl_UpgradeOps

A macro to be used by ambassador::Delegate to delegate UpgradeOps

Structs

Address

Address is the struct that defines the protocol and data payload conversion from either a public key or value

Block

BlockRegistry

A registry of open blocks (per-kernel). Think “file descriptor” table. At the moment, there’s no way to close/remove a block from this table.

BlockStat

CallResult

ExitCode

ExitCode defines the exit code from the VM invocation.

Gas

A typesafe representation of gas (internally stored as milligas).

GasTimer

A handle returned by charge_gas which must be used to mark the end of the execution associated with that gas.

MessageContext

NetworkContext

NetworkVersion

Specifies the network version

PriceList

Provides prices for operations in the VM. All costs are in milligas.

SendFlags

Flags passed to the send syscall.

TokenAmount

A quantity of native tokens. A token amount is an integer, but has a human interpretation as a value with 18 decimal places. This is a new-type in order to prevent accidental conversion from other BigInts. From/Into BigInt is missing by design.

Enums

SignatureType

Signature variants for Filecoin signatures.

Constants

RANDOMNESS_LENGTH

SECP_PUB_LEN

Secp256k1 Public key length in bytes.

SECP_SIG_LEN

Secp256k1 signature length in bytes.

SECP_SIG_MESSAGE_HASH_SIZE

Length of the signature input message hash in bytes (32).

Traits

ActorOps

Actors operations whose scope of action is actors other than the calling actor. The calling actor’s state may be consulted to resolve some.

CryptoOps

Cryptographic primitives provided by the kernel.

DebugOps

Debugging APIs.

EventOps

Eventing APIs.

IpldBlockOps

The IPLD subset of the kernel.

Kernel

The “kernel” implements the FVM interface as presented to the actors. It:

MessageOps

Accessors to query attributes of the incoming message.

NetworkOps

Network-related operations.

RandomnessOps

Randomness queries.

SelfOps

Actor state access and manipulation. Depends on BlockOps to read and write blocks in the state tree.

SendOps

The actor calling operations.

SyscallHandler

UpgradeOps

The actor upgrade operations.

Type Aliases

ActorID

Identifier for Actors, includes builtin and initialized actors

BlockId

Blocks in the block registry are addressed by an ordinal, starting from 1 (FIRST_ID). The zero value is reserved to mean “no data”, such as when actor invocations receive or return no data.

ChainEpoch

Epoch number of a chain. This acts as a proxy for time within the VM.

Cid

A Cid that contains a multihash with an allocated size of 512 bits.

MethodNum

Method number indicator for calling actor methods.

Multihash

A Multihash with the same allocated size as the Multihashes produces by this derive.

Derive Macros

Delegate

Delegate the implementation of a trait to a struct field/enum variants by adding #[derive(Delegate)] and its associated attribute #[delegate(Trait)] to it:

Multihash