Trait forest_filecoin::shim::fvm_latest::kernel::prelude::CryptoOps

pub trait CryptoOps {
    // Required methods
    fn verify_signature(
        &self,
        sig_type: SignatureType,
        signature: &[u8],
        signer: &Address,
        plaintext: &[u8]
    ) -> Result<bool, ExecutionError>;
    fn recover_secp_public_key(
        &self,
        hash: &[u8; 32],
        signature: &[u8; 65]
    ) -> Result<[u8; 65], ExecutionError>;
    fn hash(
        &self,
        code: u64,
        data: &[u8]
    ) -> Result<Multihash<64>, ExecutionError>;
}

Cryptographic primitives provided by the kernel.

Required Methods

fn verify_signature( &self, sig_type: SignatureType, signature: &[u8], signer: &Address, plaintext: &[u8] ) -> Result<bool, ExecutionError>

Verifies that a signature is valid for an address and plaintext.

fn recover_secp_public_key( &self, hash: &[u8; 32], signature: &[u8; 65] ) -> Result<[u8; 65], ExecutionError>

Given a message hash and its signature, recovers the public key of the signer.

fn hash(&self, code: u64, data: &[u8]) -> Result<Multihash<64>, ExecutionError>

Hashes input data_in using with the specified hash function, writing the output to digest_out, returning the size of the digest written to digest_out. If digest_out is to small to fit the entire digest, it will be truncated. If too large, the leftover space will not be overwritten.