Trait forest_filecoin::shim::fvm_latest::kernel::prelude::RandomnessOps

pub trait RandomnessOps {
    // Required methods
    fn get_randomness_from_tickets(
        &self,
        rand_epoch: i64
    ) -> Result<[u8; 32], ExecutionError>;
    fn get_randomness_from_beacon(
        &self,
        rand_epoch: i64
    ) -> Result<[u8; 32], ExecutionError>;
}

Randomness queries.

Required Methods

fn get_randomness_from_tickets( &self, rand_epoch: i64 ) -> Result<[u8; 32], ExecutionError>

Randomness returns a (pseudo)random byte array drawing from the latest ticket chain from a given epoch. This randomness is fork dependant but also biasable because of this.

fn get_randomness_from_beacon( &self, rand_epoch: i64 ) -> Result<[u8; 32], ExecutionError>

Randomness returns a (pseudo)random byte array drawing from the latest beacon from a given epoch. This randomness is not tied to any fork of the chain, and is unbiasable.