Trait forest_filecoin::shim::fvm_latest::kernel::prelude::SendOps

pub trait SendOps<K = Self>
where
    K: Kernel,
{
    // Required method
    fn send(
        &mut self,
        recipient: &Address,
        method: u64,
        params: u32,
        value: &TokenAmount,
        gas_limit: Option<Gas>,
        flags: SendFlags
    ) -> Result<CallResult, ExecutionError>;
}

The actor calling operations.

Required Methods

fn send( &mut self, recipient: &Address, method: u64, params: u32, value: &TokenAmount, gas_limit: Option<Gas>, flags: SendFlags ) -> Result<CallResult, ExecutionError>

Sends a message to another actor. The method type parameter K is the type of the kernel to instantiate for the receiving actor. This is necessary to support wrapping a kernel, so the outer kernel can specify its Self as the receiver’s kernel type, rather than the wrapped kernel specifying its Self. This method is part of the Kernel trait so it can refer to the Self::CallManager associated type necessary to constrain K.