Type Alias forest_filecoin::shim::fvm_latest::kernel::prelude::Cid

pub type Cid = Cid<64>;

A Cid that contains a multihash with an allocated size of 512 bits.

This is the same digest size the default multihash code table has.

If you need a CID that is generic over its digest size, use CidGeneric instead.

Aliased Type

struct Cid {
    version: Version,
    codec: u64,
    hash: Multihash<64>,
}

Fields
version: Versioncodec: u64hash: Multihash<64>