Trait forest_filecoin::shim::fvm_latest::kernel::CryptoOps

pub trait CryptoOps {
    // Required methods
    fn verify_signature(
        &self,
        sig_type: SignatureType,
        signature: &[u8],
        signer: &Address,
        plaintext: &[u8]
    ) -> Result<bool, ExecutionError>;
    fn recover_secp_public_key(
        &self,
        hash: &[u8; 32],
        signature: &[u8; 65]
    ) -> Result<[u8; 65], ExecutionError>;
    fn hash(
        &self,
        code: u64,
        data: &[u8]
    ) -> Result<Multihash<64>, ExecutionError>;
    fn compute_unsealed_sector_cid(
        &self,
        proof_type: RegisteredSealProof,
        pieces: &[PieceInfo]
    ) -> Result<Cid<64>, ExecutionError>;
    fn verify_post(
        &self,
        verify_info: &WindowPoStVerifyInfo
    ) -> Result<bool, ExecutionError>;
    fn verify_consensus_fault(
        &self,
        h1: &[u8],
        h2: &[u8],
        extra: &[u8]
    ) -> Result<Option<ConsensusFault>, ExecutionError>;
    fn batch_verify_seals(
        &self,
        vis: &[SealVerifyInfo]
    ) -> Result<Vec<bool, Global>, ExecutionError>;
    fn verify_aggregate_seals(
        &self,
        aggregate: &AggregateSealVerifyProofAndInfos
    ) -> Result<bool, ExecutionError>;
    fn verify_replica_update(
        &self,
        replica: &ReplicaUpdateInfo
    ) -> Result<bool, ExecutionError>;
}

Cryptographic primitives provided by the kernel.

Required Methods

fn verify_signature( &self, sig_type: SignatureType, signature: &[u8], signer: &Address, plaintext: &[u8] ) -> Result<bool, ExecutionError>

Verifies that a signature is valid for an address and plaintext.

fn recover_secp_public_key( &self, hash: &[u8; 32], signature: &[u8; 65] ) -> Result<[u8; 65], ExecutionError>

Given a message hash and its signature, recovers the public key of the signer.

fn hash(&self, code: u64, data: &[u8]) -> Result<Multihash<64>, ExecutionError>

Hashes input data_in using with the specified hash function, writing the output to digest_out, returning the size of the digest written to digest_out. If digest_out is to small to fit the entire digest, it will be truncated. If too large, the leftover space will not be overwritten.

fn compute_unsealed_sector_cid( &self, proof_type: RegisteredSealProof, pieces: &[PieceInfo] ) -> Result<Cid<64>, ExecutionError>

Computes an unsealed sector CID (CommD) from its constituent piece CIDs (CommPs) and sizes.

fn verify_post( &self, verify_info: &WindowPoStVerifyInfo ) -> Result<bool, ExecutionError>

Verifies a window proof of spacetime.

fn verify_consensus_fault( &self, h1: &[u8], h2: &[u8], extra: &[u8] ) -> Result<Option<ConsensusFault>, ExecutionError>

Verifies that two block headers provide proof of a consensus fault:

both headers mined by the same actor

headers are different

first header is of the same or lower epoch as the second

at least one of the headers appears in the current chain at or after epoch earliest

the headers provide evidence of a fault (see the spec for the different fault types). The parameters are all serialized block headers. The third “extra” parameter is consulted only for the “parent grinding fault”, in which case it must be the sibling of h1 (same parent tipset) and one of the blocks in the parent of h2 (i.e. h2’s grandparent). Returns nil and an error if the headers don’t prove a fault.

fn batch_verify_seals( &self, vis: &[SealVerifyInfo] ) -> Result<Vec<bool, Global>, ExecutionError>

Verifies a batch of seals. This is a privledged syscall, may only be called by the power actor during cron.

Gas: This syscall intentionally does not charge any gas (as said gas would be charged to cron). Instead, gas is pre-paid by the storage provider on pre-commit.

fn verify_aggregate_seals( &self, aggregate: &AggregateSealVerifyProofAndInfos ) -> Result<bool, ExecutionError>

Verify aggregate seals verifies an aggregated batch of prove-commits.

fn verify_replica_update( &self, replica: &ReplicaUpdateInfo ) -> Result<bool, ExecutionError>

Verify replica update verifies a snap deal: an upgrade from a CC sector to a sector with deals.