Type Alias forest_filecoin::shim::fvm_latest::kernel::BlockId

pub type BlockId = u32;

Blocks in the block registry are addressed by an ordinal, starting from 1 (FIRST_ID). The zero value is reserved to mean “no data”, such as when actor invocations receive or return no data.