Trait forest_filecoin::shim::fvm_latest::syscalls::IntoControlFlow

pub trait IntoControlFlow: Sized {
    type Value: SyscallSafe;

    // Required method
    fn into_control_flow(self) -> ControlFlow<Self::Value>;
}

The helper trait used by Syscall to convert kernel results with execution errors into results that can be handled by the wasm vm. See the documentation on Syscall for details.

Required Associated Types

type Value: SyscallSafe

Required Methods

fn into_control_flow(self) -> ControlFlow<Self::Value>

Object Safety

This trait is not object safe.