Trait forest_filecoin::shim::fvm_latest::Kernel

pub trait Kernel: ActorOps + IpldBlockOps + CircSupplyOps + CryptoOps + DebugOps + EventOps + GasOps + MessageOps + NetworkOps + RandomnessOps + SelfOps + LimiterOps + 'static {
    type CallManager: CallManager;

    // Required methods
    fn into_inner(self) -> (Self::CallManager, BlockRegistry)
       where Self: Sized;
    fn new(
        mgr: Self::CallManager,
        blocks: BlockRegistry,
        caller: u64,
        actor_id: u64,
        method: u64,
        value_received: TokenAmount,
        read_only: bool
    ) -> Self
       where Self: Sized;
    fn machine(&self) -> &<Self::CallManager as CallManager>::Machine;
    fn send<K>(
        &mut self,
        recipient: &Address,
        method: u64,
        params: u32,
        value: &TokenAmount,
        gas_limit: Option<Gas>,
        flags: SendFlags
    ) -> Result<SendResult, ExecutionError>
       where K: Kernel<CallManager = Self::CallManager>;
}

The “kernel” implements the FVM interface as presented to the actors. It:

Manages the Actor’s state.

Tracks and charges for IPLD & syscall-specific gas.

Actors may call into the kernel via the syscalls defined in the syscalls module.

Required Associated Types

type CallManager: CallManager

The Kernel’s CallManager is

Required Methods

fn into_inner(self) -> (Self::CallManager, BlockRegistry)where Self: Sized,

Consume the Kernel and return the underlying CallManager and BlockRegistry.

fn new( mgr: Self::CallManager, blocks: BlockRegistry, caller: u64, actor_id: u64, method: u64, value_received: TokenAmount, read_only: bool ) -> Selfwhere Self: Sized,

Construct a new Kernel from the given CallManager.

caller is the ID of the immediate caller.

actor_id is the ID of this actor.

method is the method that has been invoked.

value_received is value received due to the current call.

blocks is the initial block registry (should already contain the parameters).

fn machine(&self) -> &<Self::CallManager as CallManager>::Machine

The kernel’s underlying “machine”.

fn send<K>( &mut self, recipient: &Address, method: u64, params: u32, value: &TokenAmount, gas_limit: Option<Gas>, flags: SendFlags ) -> Result<SendResult, ExecutionError>where K: Kernel<CallManager = Self::CallManager>,

Sends a message to another actor. The method type parameter K is the type of the kernel to instantiate for the receiving actor. This is necessary to support wrapping a kernel, so the outer kernel can specify its Self as the receiver’s kernel type, rather than the wrapped kernel specifying its Self. This method is part of the Kernel trait so it can refer to the Self::CallManager associated type necessary to constrain K.