Module forest_filecoin::shim::fvm_shared_latest

Modules

address

bigint

chainid

clock

commcid

consensus

crypto

deal

econ

error

event

math

message

piece

randomness

receipt

reward

sector

smooth

state

sys

This module contains types exchanged at the syscall layer between actors (usually through the SDK) and the FVM.

version

Structs

DefaultNetworkParams

Params for the network. This is now continued on into mainnet and is static across networks.

Response

The outcome of a Send, covering its ExitCode and optional return data

TOTAL_FILECOIN

Total Filecoin available to the network.

ZERO_ADDRESS

Zero address used to avoid allowing it to be used for verification. This is intentionally disallowed because it is an edge case with Filecoin’s BLS signature verification.

Constants

ALLOWABLE_CLOCK_DRIFT

Allowable clock drift in validations.

BLOCKS_PER_EPOCH

The expected number of block producers in each epoch.

BLOCK_GAS_LIMIT

Total gas limit allowed per block. This is shared across networks.

HAMT_BIT_WIDTH

Default bit width for the hamt in the filecoin protocol.

IDENTITY_HASH

Multihash code for the identity hash function.

IPLD_RAW

Codec for raw data.

MAX_CID_LEN

The maximum supported CID size.

METHOD_CONSTRUCTOR

Base actor constructor method.

METHOD_SEND

Base actor send method.

TICKET_RANDOMNESS_LOOKBACK

Lookback height for retrieving ticket randomness.

TOTAL_FILECOIN_BASE

Total Filecoin supply.

WINNING_POST_SECTOR_SET_LOOKBACK

Epochs to look back for verifying PoSt proofs.

Traits

NetworkParams

Config trait which handles different network configurations.

Type Aliases

ActorID

Identifier for Actors, includes builtin and initialized actors

MethodNum

Method number indicator for calling actor methods.