Module forest_filecoin::shim::fvm_shared_latest::address

Structs

Address

Address is the struct that defines the protocol and data payload conversion from either a public key or value

DelegatedAddress

A “delegated” (f4) address.

Enums

Error

Address error

Network

Network defines the preconfigured networks to use with address encoding

Payload

Payload is the data of the Address. Variants are the supported Address protocols.

Protocol

Protocol defines the addressing protocol used to derive data to an address

Constants

BLS_PUB_LEN

BLS public key length used for validation of BLS addresses.

CHECKSUM_HASH_LEN

Length of the checksum hash for string encodings.

FIRST_NON_SINGLETON_ADDR

Defines first available ID address after builtin actors

MAX_ADDRESS_LEN

The max encoded length of an address.

MAX_SUBADDRESS_LEN

Max length of f4 sub addresses.

PAYLOAD_HASH_LEN

Hash length of payload for Secp and Actor addresses.

SECP_PUB_LEN

Uncompressed secp public key used for validation of Secp addresses.

Functions

current_network

Gets the current network.

set_current_network

Sets the default network.