Module forest_filecoin::shim::gas

Structs

Gas

GasCharge

GasChargeV3

Single gas charge in the VM. Contains information about what gas was for, as well as the amount of gas needed for computation and storage respectively.

GasChargeV4

Single gas charge in the VM. Contains information about what gas was for, as well as the amount of gas needed for computation and storage respectively.

GasCharge_latest

Single gas charge in the VM. Contains information about what gas was for, as well as the amount of gas needed for computation and storage respectively.

GasDuration

GasDurationV4

Data structure to encapsulate the optional duration which is set by the GasTimer.

GasDuration_latest

Data structure to encapsulate the optional duration which is set by the GasTimer.

GasTracker

GasV4

A typesafe representation of gas (internally stored as milligas).

Gas_latest

A typesafe representation of gas (internally stored as milligas).

PriceListV3

Provides prices for operations in the VM. All costs are in milligas.

PriceListV4

Provides prices for operations in the VM. All costs are in milligas.

Enums

PriceList

Functions

price_list_by_network_version