Struct forest_filecoin::shim::gas::PriceListV3

pub struct PriceListV3 {

    pub(crate) on_chain_message_compute: ScalingCost,
    pub(crate) on_chain_message_storage: ScalingCost,
    pub(crate) on_chain_return_compute: ScalingCost,
    pub(crate) on_chain_return_storage: ScalingCost,
    pub(crate) send_transfer_funds: Gas,
    pub(crate) send_invoke_method: Gas,
    pub(crate) address_lookup: Gas,
    pub(crate) address_assignment: Gas,
    pub(crate) actor_lookup: Gas,
    pub(crate) actor_update: Gas,
    pub(crate) actor_create_storage: Gas,
    pub(crate) sig_cost: HashMap<SignatureType, ScalingCost, RandomState>,
    pub(crate) secp256k1_recover_cost: Gas,
    pub(crate) hashing_cost: HashMap<SupportedHashes, ScalingCost, RandomState>,
    pub(crate) tipset_cid_latest: Gas,
    pub(crate) tipset_cid_historical: Gas,
    pub(crate) compute_unsealed_sector_cid_base: Gas,
    pub(crate) verify_seal_base: Gas,
    pub(crate) verify_aggregate_seal_per: HashMap<RegisteredSealProof, Gas, RandomState>,
    pub(crate) verify_aggregate_seal_steps: HashMap<RegisteredSealProof, StepCost, RandomState>,
    pub(crate) verify_post_lookup: HashMap<RegisteredPoStProof, ScalingCost, RandomState>,
    pub(crate) verify_consensus_fault: Gas,
    pub(crate) verify_replica_update: Gas,
    pub(crate) get_randomness_seed: Gas,
    pub(crate) block_memcpy: ScalingCost,
    pub(crate) block_allocate: ScalingCost,
    pub(crate) block_memory_retention_minimum: ScalingCost,
    pub(crate) block_open: ScalingCost,
    pub(crate) block_persist_storage: ScalingCost,
    pub(crate) block_persist_compute: Gas,
    pub(crate) syscall_cost: Gas,
    pub(crate) wasm_rules: WasmGasPrices,
    pub(crate) event_validation_cost: ScalingCost,
    pub(crate) event_accept_per_index_element: ScalingCost,
    pub(crate) builtin_actor_manifest_lookup: Gas,
    pub(crate) network_context: Gas,
    pub(crate) message_context: Gas,
    pub(crate) install_wasm_per_byte_cost: Gas,
    pub(crate) preloaded_actors: Vec<u64, Global>,

}

Provides prices for operations in the VM. All costs are in milligas.

Fields
on_chain_message_compute: ScalingCoston_chain_message_storage: ScalingCoston_chain_return_compute: ScalingCoston_chain_return_storage: ScalingCostsend_transfer_funds: Gassend_invoke_method: Gasaddress_lookup: Gasaddress_assignment: Gasactor_lookup: Gasactor_update: Gasactor_create_storage: Gassig_cost: HashMap<SignatureType, ScalingCost, RandomState>secp256k1_recover_cost: Gashashing_cost: HashMap<SupportedHashes, ScalingCost, RandomState>tipset_cid_latest: Gastipset_cid_historical: Gascompute_unsealed_sector_cid_base: Gasverify_seal_base: Gasverify_aggregate_seal_per: HashMap<RegisteredSealProof, Gas, RandomState>verify_aggregate_seal_steps: HashMap<RegisteredSealProof, StepCost, RandomState>verify_post_lookup: HashMap<RegisteredPoStProof, ScalingCost, RandomState>verify_consensus_fault: Gasverify_replica_update: Gasget_randomness_seed: Gasblock_memcpy: ScalingCostblock_allocate: ScalingCostblock_memory_retention_minimum: ScalingCostblock_open: ScalingCostblock_persist_storage: ScalingCostblock_persist_compute: Gassyscall_cost: Gaswasm_rules: WasmGasPricesevent_validation_cost: ScalingCostevent_accept_per_index_element: ScalingCostbuiltin_actor_manifest_lookup: Gasnetwork_context: Gasmessage_context: Gasinstall_wasm_per_byte_cost: Gaspreloaded_actors: Vec<u64, Global>
Implementations

impl PriceList

pub fn on_chain_message(&self, msg_size: usize) -> GasCharge

Returns the gas required for storing a message of a given size in the chain, plus the cost of updating the sending actor’s nonce and balance in the state-tree.

pub fn on_value_transfer(&self) -> GasCharge

Returns the gas required when invoking a method.

pub fn on_method_invocation(&self) -> GasCharge

Returns the gas required when invoking a method.

pub fn on_method_return( &self, call_depth: u32, data_size: u32 ) -> Option<GasCharge>

Returns the gas required for storing the response of a message in the chain.

pub fn on_syscall(&self) -> GasCharge

Returns the gas cost to be applied on a syscall.

pub fn on_create_actor(&self, new_address: bool) -> GasCharge

Returns the gas required for creating an actor. Pass true to when explicitly assigning a new address.

pub fn on_delete_actor(&self) -> GasCharge

Returns the gas required for deleting an actor.

pub fn on_verify_signature( &self, sig_type: SignatureType, data_len: usize ) -> GasCharge

Returns gas required for signature verification.

pub fn on_recover_secp_public_key(&self) -> GasCharge

Returns gas required for recovering signer pubkey from signature

pub fn on_hashing(&self, hasher: SupportedHashes, data_len: usize) -> GasCharge

Returns gas required for hashing data.

pub fn on_compute_unsealed_sector_cid( &self, _proof: RegisteredSealProof, _pieces: &[PieceInfo] ) -> GasCharge

Returns gas required for computing unsealed sector Cid.

pub fn on_verify_seal(&self, _info: &SealVerifyInfo) -> GasCharge

Returns gas required for seal verification.

pub fn on_verify_aggregate_seals( &self, aggregate: &AggregateSealVerifyProofAndInfos ) -> GasCharge

pub fn on_verify_replica_update( &self, _replica: &ReplicaUpdateInfo ) -> GasCharge

Returns gas required for replica verification.

pub fn on_verify_post(&self, info: &WindowPoStVerifyInfo) -> GasCharge

Returns gas required for PoSt verification.

pub fn on_verify_consensus_fault( &self, _h1_len: usize, _h2_len: usize, _extra_len: usize ) -> GasCharge

Returns gas required for verifying consensus fault.

pub fn on_get_randomness(&self, entropy_size: usize) -> GasCharge

Returns the cost of the gas required for getting randomness from the client, based on the number of bytes of entropy.

pub fn on_block_open_base(&self) -> GasCharge

Returns the base gas required for loading an object, independent of the object’s size.

pub fn on_block_open_per_byte(&self, data_size: usize) -> GasCharge

Returns the gas required for loading an object based on the size of the object.

pub fn on_block_read(&self, data_size: usize) -> GasCharge

Returns the gas required for reading a loaded object.

pub fn on_block_create(&self, data_size: usize) -> GasCharge

Returns the gas required for adding an object to the FVM cache.

pub fn on_block_link( &self, hash_code: SupportedHashes, data_size: usize ) -> GasCharge

Returns the gas required for committing an object to the state blockstore.

pub fn on_block_stat(&self) -> GasCharge

Returns the gas required for storing an object.

pub fn on_actor_lookup(&self) -> GasCharge

Returns the gas required to lookup an actor in the state-tree.

pub fn on_actor_update(&self) -> GasCharge

Returns the gas required to update an actor in the state-tree. Assumes that the actor lookup fee has already been charged.

pub fn on_actor_create(&self) -> GasCharge

Returns the gas required to create a new actor in the state-tree. Assumes that the actor lookup and update fees have already been charged.

pub fn on_self_balance(&self) -> GasCharge

Returns the gas required for accessing the balance of the current actor.

pub fn on_balance_of(&self) -> GasCharge

Returns the gas required for accessing the balance of an actor.

pub fn on_resolve_address(&self) -> GasCharge

Returns the gas required for resolving an actor address.

Might require lookup in the state tree as well as loading the state of the init actor.

pub fn on_lookup_delegated_address(&self) -> GasCharge

Returns the gas required for looking up an actor’s delegated address.

pub fn on_get_actor_code_cid(&self) -> GasCharge

Returns the gas required for getting the CID of the code of an actor.

Might require looking up the actor in the state tree.

pub fn on_get_builtin_actor_type(&self) -> GasCharge

Returns the gas required for looking up the type of a builtin actor by CID.

pub fn on_get_code_cid_for_type(&self) -> GasCharge

Returns the gas required for looking up the CID of a builtin actor by type.

pub fn on_tipset_cid(&self, lookback: bool) -> GasCharge

Returns the gas required for looking up a tipset CID with the given lookback.

pub fn on_network_context(&self) -> GasCharge

Returns the gas required for accessing the network context.

pub fn on_message_context(&self) -> GasCharge

Returns the gas required for accessing the message context.

pub fn init_memory_gas(&self, min_memory_bytes: usize) -> Gas

Returns the gas required for initializing memory.

pub fn grow_memory_gas(&self, grow_memory_bytes: usize) -> Gas

Returns the gas required for growing memory.

pub fn init_table_gas(&self, min_table_elements: u32) -> Gas

Returns the gas required for initializing tables.

pub fn on_actor_event_validate(&self, data_size: usize) -> GasCharge

pub fn on_actor_event_accept( &self, evt: &ActorEvent, serialized_len: usize ) -> GasCharge