Module forest_filecoin::shim::sector

Structs

PoStProof

RegisteredPoStProof

RegisteredSealProof

Represents a shim over RegisteredSealProof from fvm_shared with convenience methods to convert to an older version of the type

SectorInfo

Represents a shim over SectorInfo from fvm_shared with convenience methods to convert to an older version of the type

Enums

RegisteredPoStProofV3

Proof of spacetime type, indicating version and sector size of the proof.

RegisteredSealProofV3

Seal proof type which defines the version and sector size.

RegisteredSealProofV4

Seal proof type which defines the version and sector size.

SectorSize

SectorSize indicates one of a set of possible sizes in the network.

Functions

convert_window_post_proof_v1_to_v1p1

Type Aliases

SectorNumber

StoragePower

Unit of storage power (measured in bytes)