Module forest_filecoin::shim::state_tree

Structs

ActorState

Newtype to wrap different versions of fvm::state_tree::ActorState

ActorStateV2

State of all actor implementations.

ActorStateV3

State of all actor implementations.

StateRoot

State root information. Contains information about the version of the state tree, the root of the tree, and a link to the information about the tree.

StateTreeV2

State tree implementation using hamt. This structure is not threadsafe and should only be used in sync contexts.

StateTreeV3

State tree implementation using hamt. This structure is not threadsafe and should only be used in sync contexts.

discriminant 🔒

Enums

StateTree

FVM StateTree variant. The new_from_root constructor will try to resolve to a valid StateTree version or fail if we don’t support it at the moment. Other methods usage should be transparent (using shimmed versions of structures introduced in this crate::shim.

StateTreeVersion

Constants

_IMPL_NUM_FromPrimitive_FOR_StateTreeVersion 🔒

Type Definitions

ActorID

Identifier for Actors, includes builtin and initialized actors