Module forest_filecoin::state_manager

Re-exports

pub use circulating_supply::GenesisInfo;

pub use utils::is_valid_for_sending;

pub use self::errors::*;

Modules

chain_rand

circulating_supply

errors 🔒

metrics 🔒

utils

Structs

MarketBalance

External format for returning market balance from state.

StateManager

State manager handles all interactions with the internal Filecoin actors state. This encapsulates the ChainStore functionality, which only handles chain data, to allow for interactions with the underlying state of the chain. The state manager not only allows interfacing with state, but also is used when performing state transitions.

TipsetStateCache 🔒

TipsetStateCacheInner 🔒

Enums

Status 🔒

Constants

DEFAULT_TIPSET_CACHE_SIZE 🔒

NO_CALLBACK

Functions

apply_block_messages

Messages are transactions that produce new states. The state (usually referred to as the ‘state-tree’) is a mapping from actor addresses to actor states. Each block contains the hash of the state-tree that should be used as the starting state when executing the block messages.

validate_tipsets

Type Aliases

CidPair 🔒

Intermediary for retrieving state objects and updating actor states.