Function forest_filecoin::state_manager::apply_block_messages

pub fn apply_block_messages<DB, CB>(
    genesis_timestamp: u64,
    chain_index: Arc<ChainIndex<Arc<DB>>>,
    chain_config: Arc<ChainConfig>,
    beacon: Arc<BeaconSchedule>,
    engine: &MultiEngine,
    tipset: Arc<Tipset>,
    callback: Option<CB>
) -> Result<(Cid, Cid), Error>where
    DB: Blockstore + Send + Sync + 'static,
    CB: FnMut(&Cid, &ChainMessage, &ApplyRet) -> Result<(), Error>,

Messages are transactions that produce new states. The state (usually referred to as the ‘state-tree’) is a mapping from actor addresses to actor states. Each block contains the hash of the state-tree that should be used as the starting state when executing the block messages.

Execution environment

Transaction execution has the following inputs:

a current state-tree (stored as IPLD in a key-value database). This reference is in Tipset::parent_state.

up to 900 past state-trees. See https://docs.filecoin.io/reference/general/glossary/#finality.

up to 900 past tipset IDs.

a deterministic source of randomness.

the circulating supply of FIL (see https://filecoin.io/blog/filecoin-circulating-supply/). The circulating supply is determined by the epoch and the states of a few key actors.

the base fee (see https://spec.filecoin.io/systems/filecoin_vm/gas_fee/). This value is defined by tipset.parent_base_fee.

the genesis timestamp (UNIX epoch time when the first block was mined/created).

a chain configuration (maps epoch to network version, has chain specific settings).

The result of running a set of block messages is an index to the final state-tree and an index to an array of message receipts (listing gas used, return codes, etc).

Cron and null tipsets

Once per epoch, after all messages have run, a special ‘cron’ transaction must be executed. The tasks of the ‘cron’ transaction include running batch jobs and keeping the state up-to-date with the current epoch.

It can happen that no blocks are mined in an epoch. The tipset for such an epoch is called a null tipset. A null tipset has no identity and cannot be directly executed. This is a problem for ‘cron’ which must run for every epoch, even if there are no messages. The fix is to run ‘cron’ if there are any null tipsets between the current epoch and the parent epoch.

Imagine the blockchain looks like this with a null tipset at epoch 9:

┌────────┐ ┌────┐ ┌───────┐  ┌───────┐
│Epoch 10│ │Null│ │Epoch 8├──►Epoch 7├─►
└───┬────┘ └────┘ └───▲───┘  └───────┘
    └─────────────────┘

The parent of tipset-epoch-10 is tipset-epoch-8. Before executing the messages in epoch 10, we have to run cron for epoch 9. However, running ‘cron’ requires the timestamp of the youngest block in the tipset (which doesn’t exist because there are no blocks in the tipset). Lotus dictates that the timestamp of a null tipset is 30s * epoch after the genesis timestamp. So, in the above example, if the genesis block was mined at time X, the null tipset for epoch 9 will have timestamp X + 30 * 9.

Migrations

Migrations happen between network upgrades and modify the state tree. If a migration is scheduled for epoch 10, it will be run after the messages for epoch 10. The tipset for epoch 11 will link the state-tree produced by the migration.

Example timeline with a migration at epoch 10:

Tipset-epoch-10 executes, producing state-tree A.

Migration consumes state-tree A and produces state-tree B.

Tipset-epoch-11 executes, consuming state-tree B (rather than A).

Note: The migration actually happens when tipset-epoch-11 executes. This is because tipset-epoch-10 may be null and therefore not executed at all.

Caching

Scanning the blockchain to find past tipsets and state-trees may be slow. The ChainStore caches recent tipsets to make these scans faster.