Module forest_filecoin::state_migration

Modules

common 🔒

Common code that’s shared across all migration code. Each network upgrade / state migration code lives in their own module.

nv17 🔒

This module contains the migration logic for the NV17 upgrade. The corresponding Go implementation can be found here: https://github.com/filecoin-project/go-state-types/blob/master/builtin/v9/migration/top.go

nv18 🔒

This module contains the migration logic for the NV18 upgrade. The corresponding Go implementation can be found here: https://github.com/filecoin-project/go-state-types/blob/master/builtin/v10/migration/top.go

nv19 🔒

This module contains the migration logic for the NV19 upgrade. The corresponding Go implementation can be found here: https://github.com/filecoin-project/go-state-types/blob/master/builtin/v11/migration/top.go

nv21 🔒

This module contains the migration logic for the NV21 upgrade. The corresponding Go implementation can be found here: https://github.com/filecoin-project/go-state-types/blob/master/builtin/v12/migration/top.go

nv21fix 🔒

This module contains the migration logic for the NV21 calibration network fix.

nv21fix2 🔒

This module contains the migration logic for the NV21 calibration network fix.

nv22 🔒

This module contains the migration logic for the NV22 upgrade. The corresponding Go implementation can be found here: https://github.com/filecoin-project/go-state-types/blob/master/builtin/v12/migration/top.go

type_migrations 🔒

Functions

run_state_migrations

Run state migrations

Type Aliases

RunMigration 🔒