Module forest_filecoin::state_migration::common

Common code that’s shared across all migration code. Each network upgrade / state migration code lives in their own module.

Modules

macros 🔒

migration_job 🔒

migrators 🔒

state_migration 🔒

verifier 🔒

Structs

ActorMigrationInput 🔒

ActorMigrationOutput 🔒

Output of actor migration job.

MigrationCache 🔒

Cache of existing CID to CID migrations for an actor.

TypeMigrator 🔒

Type that implements TypeMigration for different type pairs. Prefer using a single struct so that the compiler could catch duplicate implementations

Traits

ActorMigration 🔒

Trait that defines the interface for actor migration job.

PostMigrationCheck 🔒

Trait defining the interface for actor migration verifier.

PostMigrator 🔒

Trait that defines the interface for actor migration job to be executed after the state migration.

TypeMigration 🔒

Trait that migrates from one data structure to another, similar to std::convert::TryInto trait but taking an extra block store parameter

Type Aliases

Migrator 🔒

PostMigrationCheckArc 🔒

Sized wrapper of PostMigrationCheck.

PostMigratorArc 🔒

Sized wrapper of PostMigrator.