Trait forest_filecoin::state_migration::common::ActorMigration

pub(super) trait ActorMigration<BS: Blockstore> {
    // Required method
    fn migrate_state(
        &self,
        store: &BS,
        input: ActorMigrationInput
    ) -> Result<Option<ActorMigrationOutput>>;

    // Provided method
    fn is_deferred(&self) -> bool { ... }
}

Trait that defines the interface for actor migration job.

Required Methods

fn migrate_state( &self, store: &BS, input: ActorMigrationInput ) -> Result<Option<ActorMigrationOutput>>

Provided Methods

fn is_deferred(&self) -> bool

Some migration jobs might need to be deferred to be executed after the regular state migration. These may require some metadata collected during other migrations.