Trait forest_filecoin::state_migration::common::TypeMigration

pub(super) trait TypeMigration<From, To> {
    // Required method
    fn migrate_type(from: From, store: &impl Blockstore) -> Result<To>;
}

Trait that migrates from one data structure to another, similar to std::convert::TryInto trait but taking an extra block store parameter

Required Methods

fn migrate_type(from: From, store: &impl Blockstore) -> Result<To>