Function forest_filecoin::tool::subcommands::api_cmd::compare_apis

async fn compare_apis(
    forest: ApiInfo,
    lotus: ApiInfo,
    snapshot_files: Vec<PathBuf>,
    config: ApiTestFlags
) -> Result<()>

Compare two RPC providers. The providers are labeled forest and lotus, but other nodes may be used (such as venus). The lotus node is assumed to be correct and the forest node will be marked as incorrect if it deviates.

If snapshot files are provided, these files will be used to generate additional tests.

Example output:

| RPC Method                        | Forest              | Lotus         |
|-----------------------------------|---------------------|---------------|
| Filecoin.ChainGetBlock            | Valid               | Valid         |
| Filecoin.ChainGetGenesis          | Valid               | Valid         |
| Filecoin.ChainGetMessage (67)     | InternalServerError | Valid         |

The number after a method name indicates how many times an RPC call was tested.