Module forest_filecoin::utils::encoding

Re-exports

pub use cid_de_cbor::extract_cids;

Modules

cid_de_cbor 🔒

de

Generic data structure deserialization framework.

fallback_de_ipld_dagcbor 🔒

ser

Generic data structure serialization framework.

serde_byte_array

serde_bytes with max length check

uvibytes

Traits

Deserializer

A data format that can deserialize any data structure supported by Serde.

Serializer

A data format that can serialize any data structure supported by Serde.

Functions

blake2b_256

Generates BLAKE2b hash of fixed 32 bytes size.

from_slice_with_fallback

This method will attempt to de-serialize given bytes using the regular serde_ipld_dagcbor::from_slice. Due to a historical issue in Lotus (see more in FIP-0027, we must still support strings with invalid UTF-8 bytes. On a failure, it will retry the operation using the fallback that will de-serialize strings with invalid UTF-8 bytes as bytes.

prover_id_from_u64