Module forest_filecoin::utils::proofs_api

Re-exports

pub use parameters::set_proofs_parameter_cache_dir_env;

pub use paramfetch::ensure_params_downloaded;

pub use paramfetch::get_params_default;

pub use paramfetch::SectorSizeOpt;

Modules

parameters 🔒

This module contains the logic for storing and verifying the proofs parameters.

paramfetch 🔒

This module contains the logic for fetching the proofs parameters from the network. As a general rule, the parameters are first fetched from ChainSafe’s Cloudflare R2 bucket, if that fails (or is overridden by PROOFS_ONLY_IPFS_GATEWAY_ENV), the IPFS gateway is used as a fallback.

Functions

is_env_truthy 🔒

Check if the given environment variable is set to truthy value.