Function forest_filecoin::utils::proofs_api::parameters::set_proofs_parameter_cache_dir_env

pub fn set_proofs_parameter_cache_dir_env(data_dir: &Path)

Forest uses a set of external crates for verifying the proofs generated by the miners. These external crates require a specific set of parameter files to be located at in a specific folder. By default, it is /var/tmp/filecoin-proof-parameters but it can be overridden by the FIL_PROOFS_PARAMETER_CACHE environment variable. Forest will automatically download the parameter files from Cloudflare/IPFS and verify their validity. For consistency, Forest will prefer to download the files it’s local data directory. To this end, the FIL_PROOFS_PARAMETER_CACHE environment variable is updated before the parameters are downloaded.

More information available here.