Module forest_filecoin::utils::proofs_api::paramfetch

Structs

ParameterData 🔒

Enums

SectorSizeOpt

Sector size options for fetching.

Constants

DEFAULT_PARAMETERS 🔒

DIR_ENV 🔒

GATEWAY 🔒

GATEWAY_ENV 🔒

PARAM_DIR 🔒

TRUST_PARAMS_ENV 🔒

Functions

check_file 🔒

ensure_params_downloaded

Ensures the parameter files are downloaded to cache dir

fetch_params 🔒

fetch_verify_params 🔒

get_params

Get proofs parameters and all verification keys for a given sector size given a parameter JSON manifest.

get_params_default

Get proofs parameters and all verification keys for a given sector size using default manifest.

param_dir 🔒

set_proofs_parameter_cache_dir_env

Forest uses a set of external crates for verifying the proofs generated by the miners. These external crates require a specific set of parameter files to be located at in a specific folder. By default, it is /var/tmp/filecoin-proof-parameters but it can be overridden by the FIL_PROOFS_PARAMETER_CACHE environment variable. Forest will automatically download the parameter files from IPFS and verify their validity. For consistency, Forest will prefer to download the files it’s local data directory. To this end, the FIL_PROOFS_PARAMETER_CACHE environment variable is updated before the parameters are downloaded.

Type Aliases

ParameterMap 🔒