Crate forest_filecoin

Modules

auth ğŸ”’

beacon ğŸ”’

blocks ğŸ”’

chain ğŸ”’

chain_sync ğŸ”’

cid_collections ğŸ”’

cli ğŸ”’

cli_shared ğŸ”’

daemon ğŸ”’

db ğŸ”’

fil_cns ğŸ”’

genesis ğŸ”’

interpreter ğŸ”’

ipld ğŸ”’

key_management ğŸ”’

libp2p ğŸ”’

libp2p_bitswap ğŸ”’

Another libp2p bitswap(SPEC) implementation in Rust.

lotus_json ğŸ”’

In the Filecoin ecosystem, there are TWO different ways to present a domain object:

message ğŸ”’

message_pool ğŸ”’

metrics ğŸ”’

networks ğŸ”’

rpc ğŸ”’

rpc_api ğŸ”’

In general, forest wants to support the same RPC messages as lotus (go implementation of Filecoin).

rpc_client ğŸ”’

shim ğŸ”’

state_manager ğŸ”’

state_migration ğŸ”’

statediff ğŸ”’

tool ğŸ”’

utils ğŸ”’

wallet ğŸ”’

Macros

define_system_states

Define type aliases for system actor State types before and after the state migration, namely SystemStateOld and SystemStateNew

impl_system

Implements fn system_migrator, requiring proper system actor State types being defined by define_system_states macro.

impl_verifier

Implements Verifier, requiring proper Manifest types and system actor State types being defined by define_manifests and define_system_states macros respectively.

Structs

Client

Config

KeyStore

KeyStore structure, this contains a set of KeyInfos indexed by address.

Enums

KeyStoreConfig

Constants

ENCRYPTED_KEYSTORE_NAME

FOREST_KEYSTORE_PHRASE_ENV

Environmental variable which holds the KeyStore encryption phrase.

JWT_IDENTIFIER

constant string that is used to identify the JWT secret key in KeyStore

KEYSTORE_NAME

Statics

GLOBAL ğŸ”’

Functions

forest_main

forest_tool_main

forest_wallet_main

forestd_main

verify_token

Verify JWT Token and return the allowed permissions from token