Module forest_filecoin::blocks::tipset

Modules

lotus_json

Tipset isn’t just plain old data - it has an invariant (all block headers are valid) So there is custom de-serialization here

Structs

FullTipset

FullTipset is an expanded version of a tipset that contains all the blocks and messages.

Tipset

An immutable set of blocks at the same height with the same parent set. Blocks in a tipset are canonically ordered by ticket size.

TipsetKey

A set of CIDs forming a unique key for a Tipset. Equal keys will have equivalent iteration order, but note that the CIDs are not maintained in the same order as the canonical iteration order of blocks in a tipset (which is by ticket)

Enums

CreateTipsetError

Constants

__CFG_VIS_ACCUMULATOR_MUST_EXPAND_ONCE_OTHERWISE_IS_A_BUG_2633470583269875123 🔒

Functions

verify_block_headers 🔒