Module forest_filecoin::cli::subcommands

Modules

auth_cmd ğŸ”’

chain_cmd ğŸ”’

config_cmd ğŸ”’

info_cmd ğŸ”’

mpool_cmd ğŸ”’

net_cmd ğŸ”’

send_cmd ğŸ”’

shutdown_cmd ğŸ”’

snapshot_cmd ğŸ”’

state_cmd ğŸ”’

sync_cmd ğŸ”’

Structs

Cli

CLI structure generated when interacting with Forest binary

Enums

Subcommand

Forest binary sub-commands available.

Functions

cli_error_and_die

Print an error message and exit the program with an error code Used for handling high level errors such as invalid parameters

format_vec_pretty ğŸ”’

Format a vector to a prettified string

print_pretty_lotus_json ğŸ”’

Prints a pretty HTTP JSON-RPC response result

print_rpc_res_bytes ğŸ”’

Prints a bytes HTTP JSON-RPC response result

print_rpc_res_cids ğŸ”’

Prints a tipset from a HTTP JSON-RPC response result

prompt_confirm