Module forest_filecoin::daemon

Modules

bundle

db_util 🔒

main

Statics

IPC_PATH 🔒

Functions

asyncify 🔒

Run the closure on a thread where blocking is allowed

create_password 🔒

Loops until the user provides two matching passwords.

get_actual_chain_name

handle_admin_token 🔒

Generates, prints and optionally writes to a file the administrator JWT token.

input_password_to_load_encrypted_keystore 🔒

Prompts for password, looping until the KeyStore is successfully loaded.

ipc_shmem_conf

load_or_create_keystore 🔒

This may:

maybe_increase_fd_limit 🔒

Increase the file descriptor limit to a reasonable number. This prevents the node from failing if the default soft limit is too low. Note that the value is only increased, never decreased.

propagate_error 🔒

returns the first error with which any of the services end, or never returns at all

set_snapshot_path_if_needed 🔒

If our current chain is below a supported height, we need a snapshot to bring it up to a supported height. If we’ve not been given a snapshot by the user, get one.

start 🔒

Starts daemon process

start_interruptable

unblock_parent_process 🔒