Module forest_filecoin::health::endpoints

Structs

MessageAccumulator 🔒

Sample message accumulator for healthcheck responses. It is intended to accumulate messages for verbose responses.

Constants

VERBOSE_PARAM 🔒

Query parameter for verbose responses

Functions

check_epoch_up_to_date 🔒

Checks if the current epoch of the node is not too far behind the network. Making the threshold too strict can cause the node to repeatedly report as not ready, especially in case of forking.

check_peers_connected 🔒

check_rpc_server_running 🔒

check_sync_state_complete 🔒

check_sync_state_not_error 🔒

healthz 🔒

This endpoint is a combination of the [livez] and [readyz] endpoints, except that the node doesn’t have to be fully synced. Deprecated in the Kubernetes world, but still used in some setups.

livez 🔒

Liveness probes determine whether or not an application running in a container is in a healthy state. The idea behind a liveness probe is that it fails for prolonged period of time, then the application should be restarted. In our case, we require:

readyz 🔒

Readiness probes determine whether or not a container is ready to serve requests. The goal is to determine if the application is fully prepared to accept traffic. In our case, we require: