Module forest_filecoin::libp2p::peer_manager

Structs

PeerInfo 🔒

Contains info about the peer’s head Tipset, as well as the request stats.

PeerManager

Thread safe peer manager which handles peer management for the ChainExchange protocol.

PeerSets 🔒

Peer tracking sets, these are handled together to avoid race conditions or deadlocks when updating state.

Enums

PeerOperation

Constants

GLOBAL_INV_ALPHA 🔒

Global duration multiplier, affects duration delta change.

LOCAL_INV_ALPHA 🔒

Local duration multiplier, affects duration delta change.

NEW_PEER_MUL 🔒

New peer multiplier slightly less than 1 to incentivize choosing new peers.

SHUFFLE_PEERS_PREFIX 🔒

Defines max number of peers to send each chain exchange request to.

Functions

log_time 🔒

remove_peer 🔒