Trait forest_filecoin::lotus_json::HasLotusJson

pub trait HasLotusJson: Sized {
    type LotusJson: Serialize + DeserializeOwned;

    // Required methods
    fn snapshots() -> Vec<(Value, Self)>;
    fn into_lotus_json(self) -> Self::LotusJson;
    fn from_lotus_json(lotus_json: Self::LotusJson) -> Self;
}

Required Associated Types

type LotusJson: Serialize + DeserializeOwned

The struct representing JSON. You should #[derive(Deserialize, Serialize)] on it.

Required Methods

fn snapshots() -> Vec<(Value, Self)>

To ensure code quality, conversion to/from lotus JSON MUST be tested. Provide snapshots of the JSON, and the domain type it should serialize to.

Serialization and de-serialization of the domain type should match the snapshot.

If using decl_and_test, this test is automatically run for you, but if the test is out-of-module, you must call assert_all_snapshots manually.

fn into_lotus_json(self) -> Self::LotusJson

fn from_lotus_json(lotus_json: Self::LotusJson) -> Self

Object Safety

This trait is not object safe.