Module forest_filecoin::message_pool::msgpool

Modules

metrics 🔒

msg_pool 🔒

provider 🔒

selection

Contains routines for message selection APIs. Whenever a miner is ready to create a block for a tipset, it invokes the select_messages API which selects an appropriate set of messages such that it optimizes miner reward and chain capacity. See https://docs.filecoin.io/mine/lotus/message-pool/#message-selection for more details

utils 🔒

Constants

BASE_FEE_LOWER_BOUND_FACTOR 🔒

BASE_FEE_LOWER_BOUND_FACTOR_CONSERVATIVE 🔒

MIN_GAS 🔒

RBF_DENOM 🔒

RBF_NUM 🔒

REPLACE_BY_FEE_RATIO 🔒

REPUB_MSG_LIMIT 🔒

Functions

add_to_selected_msgs 🔒

This is a helper function for head_change. This method will add a signed message to the given messages selected by priority HashMap.

get_state_sequence 🔒

Get the state of the base_sequence for a given address in the current Tipset

head_change

This function will revert and/or apply tipsets to the message pool. This function should be called every time that there is a head change in the message pool.

remove_from_selected_msgs 🔒

This is a helper function for head_change. This method will remove a sequence for a from address from the messages selected by priority hash-map. It also removes the ‘from’ address and sequence from the MessagePool.

republish_pending_messages 🔒

select_messages_for_block 🔒

Select messages from the mempool to be included in the next block that builds on a given base tipset. The messages should be eligible for inclusion based on their sequences and the overall number of them should observe block gas limits.