Module forest_filecoin::rpc::methods::state

Re-exports

pub use types::*;

Modules

types 🔒

Macros

for_each_method 🔒

Enums

MinerGetBaseInfo

MsigGetAvailableBalance

MsigGetPending

StateAccountKey

gets the public key address of the given ID address See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateAccountKey

StateCall

StateCirculatingSupply

StateDealProviderCollateralBounds

StateFetchRoot

Traverse an IPLD directed acyclic graph and use libp2p-bitswap to request any missing nodes. This function has two primary uses: (1) Downloading specific state-roots when Forest deviates from the mainline blockchain, (2) fetching historical state-trees to verify past versions of the consensus rules.

StateGetActor

StateGetBeaconEntry

StateGetRandomnessFromBeacon

Get randomness from beacon

StateGetRandomnessFromTickets

Get randomness from tickets

StateGetReceipt

returns the message receipt for the given message

StateListMessages

Looks back and returns all messages with a matching to or from address, stopping at the given height.

StateListMiners

StateLookupID

retrieves the ID address of the given address See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateLookupID

StateMarketBalance

looks up the Escrow and Locked balances of the given address in the Storage Market

StateMarketDeals

StateMarketStorageDeal

StateMinerActiveSectors

StateMinerAvailableBalance

StateMinerDeadlines

StateMinerFaults

looks up the miner power of the given address.

StateMinerInfo

looks up the miner info of the given address.

StateMinerInitialPledgeCollateral

StateMinerPartitions

Return all partitions in the specified deadline

StateMinerPower

looks up the miner power of the given address.

StateMinerProvingDeadline

StateMinerRecoveries

StateMinerSectorCount

Returns the number of sectors in a miner’s sector set and proving set

StateMinerSectors

StateNetworkName

StateNetworkVersion

StateReadState

Get read state

StateReplay

StateSearchMsg

Searches for a message in the chain, and returns its receipt and the tipset where it was executed. See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateSearchMsg

StateSearchMsgLimited

Looks back up to limit epochs in the chain for a message, and returns its receipt and the tipset where it was executed. See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateSearchMsgLimited

StateSectorGetInfo

StateSectorPreCommitInfo

StateVMCirculatingSupplyInternal

StateVerifiedClientStatus

StateWaitMsg

looks back in the chain for a message. If not found, it blocks until the message arrives on chain, and gets to the indicated confidence depth.

Constants

INITIAL_PLEDGE_DEN 🔒

INITIAL_PLEDGE_NUM 🔒

STATE_DECODE_PARAMS

Functions

lock_pop 🔒