Module forest_filecoin::rpc::reflect

Forest wishes to provide OpenRPC definitions for Filecoin APIs. To do this, it needs:

JSON Schema definitions for all the argument and return types.

The number of arguments (arity) and names of those arguments for each RPC method.

As a secondary objective, we wish to provide an RPC client for our CLI, and internal tests against Lotus.

The RpcMethod trait encapsulates all the above at a single site.

schemars::JsonSchema provides schema definitions,

RpcMethod defining arity and actually dispatching the function calls.

SelfDescribingRpcModule actually does the work to create the OpenRPC document.

Modules

jsonrpc_types

A transcription of types from the JSON-RPC 2.0 Specification.

openrpc_types

A transcription of types from the OpenRPC Specification.

parser 🔒

util 🔒

Macros

do_impls 🔒

Structs

SelfDescribingRpcModule

Enums

ApiVersion

Lotus groups methods into API versions.

ConcreteCallingConvention

openrpc_types::ParamStructure describes accepted param format. This is an actual param format, used to decide how to construct arguments.

Traits

Params

A tuple of ARITY arguments.

RpcMethod

A definition of an RPC method handler which can be registered with a SelfDescribingRpcModule.

RpcMethodExt

Utility methods, defined as an extension trait to avoid having to specify ARITY in user code.

Functions

params2params 🔒

unexpected 🔒

Type Aliases

Ctx

Type to be used by RpcMethod::handle.