Trait forest_filecoin::rpc::reflect::RpcMethod

pub trait RpcMethod<const ARITY: usize> {
    type Params: Params<ARITY>;
    type Ok: HasLotusJson;

    const NAME: &'static str;
    const PARAM_NAMES: [&'static str; ARITY];
    const API_VERSION: ApiVersion;
    const PERMISSION: Permission;
    const TAGS: &'static [Tag] = _;
    const SUMMARY: Option<&'static str> = None;
    const DESCRIPTION: Option<&'static str> = None;

    // Required method
    fn handle(
        ctx: Arc<RPCState<impl Blockstore + Send + Sync + 'static>>,
        params: Self::Params
    ) -> impl Future<Output = Result<Self::Ok, Error>> + Send;
}

A definition of an RPC method handler which:

can be registered with an RpcModule.

can describe itself in OpenRPC.

Note, an earlier draft of this trait had an additional type parameter for Ctx for generality. However, fixing it as [Ctx<...>] saves on complexity/confusion for implementors, at the expense of handler flexibility, which could come back to bite us.

All handlers accept the same type.

All Ctxs must be Send + Sync + 'static due to bounds on RpcModule.

Handlers don’t specialize on top of the given bounds, but they MAY relax them.

Required Associated Types

type Params: Params<ARITY>

Types of each argument. Option-al arguments MUST follow mandatory ones.

type Ok: HasLotusJson

Return value of this method.

Required Associated Constants

const NAME: &'static str

Method name.

const PARAM_NAMES: [&'static str; ARITY]

Name of each argument, MUST be unique.

const API_VERSION: ApiVersion

See ApiVersion.

const PERMISSION: Permission

See Permission

Provided Associated Constants

const TAGS: &'static [Tag] = _

See Tag.

const SUMMARY: Option<&'static str> = None

Becomes openrpc_types::Method::summary.

const DESCRIPTION: Option<&'static str> = None

Becomes openrpc_types::Method::description.

Required Methods

fn handle( ctx: Arc<RPCState<impl Blockstore + Send + Sync + 'static>>, params: Self::Params ) -> impl Future<Output = Result<Self::Ok, Error>> + Send

Logic for this method.

Object Safety

This trait is not object safe.