Module forest_filecoin::rpc::state_api

Functions

lock_pop 🔒

miner_get_base_info

msig_get_available_balance

msig_get_pending

state_account_key

gets the public key address of the given ID address See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateAccountKey

state_call

runs the given message and returns its result without any persisted changes.

state_circulating_supply

state_fetch_root

Traverse an IPLD directed acyclic graph and use libp2p-bitswap to request any missing nodes. This function has two primary uses: (1) Downloading specific state-roots when Forest deviates from the mainline blockchain, (2) fetching historical state-trees to verify past versions of the consensus rules.

state_get_actor 🔒

state_get_network_version

state_get_randomness_from_beacon

Get randomness from beacon

state_get_randomness_from_tickets

Get randomness from tickets

state_get_receipt

returns the message receipt for the given message

state_list_messages 🔒

Looks back and returns all messages with a matching to or from address, stopping at the given height.

state_list_miners

state_lookup_id

retrieves the ID address of the given address See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateLookupID

state_market_balance

looks up the Escrow and Locked balances of the given address in the Storage Market

state_market_deals

state_miner_active_sectors

state_miner_deadlines

state_miner_faults

looks up the miner power of the given address.

state_miner_info

looks up the miner info of the given address.

state_miner_power

looks up the miner power of the given address.

state_miner_proving_deadline

state_miner_recoveries

state_miner_sector_count

state_network_name

gets network name from state manager

state_read_state

Get read state

state_replay

returns the result of executing the indicated message, assuming it was executed in the indicated tipset.

state_search_msg

Searches for a message in the chain, and returns its receipt and the tipset where it was executed. See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateSearchMsg

state_search_msg_limited

Looks back up to limit epochs in the chain for a message, and returns its receipt and the tipset where it was executed. See https://github.com/filecoin-project/lotus/blob/master/documentation/en/api-v0-methods.md#StateSearchMsgLimited

state_sector_get_info

Get state sector info using sector no

state_verified_client_status 🔒

state_vm_circulating_supply_internal 🔒

state_wait_msg

looks back in the chain for a message. If not found, it blocks until the message arrives on chain, and gets to the indicated confidence depth.

Type Aliases

RandomnessParams 🔒