Module forest_filecoin::rpc::types

Types that are shared between APIs.

If a type here is used by only one API, it should be relocated.

Modules

address_impl 🔒

beneficiary_impl 🔒

deal_impl 🔒

miner_impl 🔒

sector_impl 🔒

tsk_impl 🔒

Structs

AddressOrEmpty

This wrapper is needed because of a bug in Lotus. See: https://github.com/filecoin-project/lotus/issues/11461.

ApiActorState

ApiDeadline

ApiDeadlineInfo

ApiDealProposal

ApiDealState

ApiMarketDeal

ApiState

ApiTipsetKey

BeneficiaryTermLotusJson

CirculatingSupply

DealCollateralBounds

MarketDeal

MessageFilter

MessageLookup

MessageSendSpec

MinerInfoLotusJson

MinerPartitions

MinerPowerLotusJson

MinerSectors

PeerID

PendingBeneficiaryChangeLotusJson

SectorOnChainInfo

SectorPreCommitInfo

SectorPreCommitOnChainInfo

Transaction