Module forest_filecoin::shim::fvm_latest

Modules

call_manager

engine

executor

externs

This module contains the logic to invoke the node by traversing Boundary A.

gas

kernel

machine

state_tree

syscalls

trace

Macros

ambassador_impl_ActorOps

A macro to be used by ambassador::Delegate to delegate ActorOps

ambassador_impl_ActorOps_body_enum

ambassador_impl_ActorOps_body_single_struct

ambassador_impl_CryptoOps

A macro to be used by ambassador::Delegate to delegate CryptoOps

ambassador_impl_CryptoOps_body_enum

ambassador_impl_CryptoOps_body_single_struct

ambassador_impl_DebugOps

A macro to be used by ambassador::Delegate to delegate DebugOps

ambassador_impl_DebugOps_body_enum

ambassador_impl_DebugOps_body_single_struct

ambassador_impl_EventOps

A macro to be used by ambassador::Delegate to delegate EventOps

ambassador_impl_EventOps_body_enum

ambassador_impl_EventOps_body_single_struct

ambassador_impl_IpldBlockOps

A macro to be used by ambassador::Delegate to delegate IpldBlockOps

ambassador_impl_IpldBlockOps_body_enum

ambassador_impl_IpldBlockOps_body_single_struct

ambassador_impl_MessageOps

A macro to be used by ambassador::Delegate to delegate MessageOps

ambassador_impl_MessageOps_body_enum

ambassador_impl_MessageOps_body_single_struct

ambassador_impl_NetworkOps

A macro to be used by ambassador::Delegate to delegate NetworkOps

ambassador_impl_NetworkOps_body_enum

ambassador_impl_NetworkOps_body_single_struct

ambassador_impl_RandomnessOps

A macro to be used by ambassador::Delegate to delegate RandomnessOps

ambassador_impl_RandomnessOps_body_enum

ambassador_impl_RandomnessOps_body_single_struct

ambassador_impl_SelfOps

A macro to be used by ambassador::Delegate to delegate SelfOps

ambassador_impl_SelfOps_body_enum

ambassador_impl_SelfOps_body_single_struct

ambassador_impl_SendOps

A macro to be used by ambassador::Delegate to delegate SendOps

ambassador_impl_SendOps_body_enum

ambassador_impl_SendOps_body_single_struct

ambassador_impl_UpgradeOps

A macro to be used by ambassador::Delegate to delegate UpgradeOps

ambassador_impl_UpgradeOps_body_enum

ambassador_impl_UpgradeOps_body_single_struct

syscall_error

Convenience macro for generating Actor Errors

total_enum_map

Create a mapping from enum items to values in a way that guarantees at compile time that we did not miss any member, in any of the prices, even if the enum gets a new member later.

Structs

DefaultKernel

The “default” Kernel implementation.

Traits

Kernel

The “kernel” implements the FVM interface as presented to the actors. It: